Update header and footer styles
