+++ apps/web/app/components/Header.tsx
@@ -1,0 +1,158 @@
+import Link from "next/link";
+import ds from "design-system/";
+import NotificationsLink from "../[locale]/notifications/NotificationsLink";
+import UserIcon from "./UserIcon";
+import { useTranslations } from "next-intl";
+import { headers } from "next/headers";
+import "./Header.css";
+export default () => {
+  const t = useTranslations("Header");
+  const pathSlice = headers().get("x-pathname")?.split("/") ?? [];
+  const path = pathSlice.slice(2)?.join("/") || "";
+
+  return (
+    <header role="banner" data-module="govie-header" className="govie-header">
+      <div
+        className="govie-header__container govie-width-container"
+        // all designs are made for 1440px
+        style={{
+          maxWidth: "1440px",
+          width: "100%",
+          display: "flex",
+          alignItems: "center",
+          height: "80px",
+          boxSizing: "border-box",
+        }}
+      >
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "space-between",
+            alignItems: "center",
+            width: "100%",
+          }}
+        >
+          <div style={{ display: "flex", alignItems: "center" }}>
+            <div style={{ paddingRight: "15px" }}>
+              <a
+                href="/"
+                className="govie-header__link govie-header__link--homepage"
+                style={{ display: "block" }}
+              >
+                <svg
+                  width="110"
+                  height="40"
+                  viewBox="0 0 110 40"
+                  fill="none"
+                  xmlns="http://www.w3.org/2000/svg"
+                  aria-hidden="true"
+                  focusable="false"
+                  style={{ verticalAlign: "bottom" }}
+                >
+                  <path
+                    d="M28.9872 0.00119403C28.8652 6.41911 28.8 13.0161 28.8 19.7552C28.8 26.4943 28.8652 33.0913 28.9872 39.5093C29.0398 39.5093 29.0918 39.5104 29.1444 39.5104C29.197 39.5104 29.2491 39.5093 29.3017 39.5093C29.4236 33.0913 29.4888 26.4943 29.4888 19.7552C29.4888 13.0161 29.4236 6.41911 29.3017 0.00119403C29.2491 0.00119403 29.197 0 29.1444 0C29.0918 0 29.0398 0.000597015 28.9872 0.00119403Z"
+                    fill="currentColor"
+                  ></path>
+                  <path
+                    fillRule="evenodd"
+                    clipRule="evenodd"
+                    d="M50.3984 14.326C50.7099 15.3224 50.8043 15.7278 50.8043 16.4442C50.8043 19.5588 48.1844 21.7075 44.3789 21.7075C43.5054 21.7075 43.0068 21.6454 41.7901 21.3964C41.4158 21.8633 41.2287 22.2997 41.2287 22.7045C41.2287 23.483 41.9461 23.8257 43.7242 23.9815L46.4691 24.1994C48.8397 24.3863 49.089 24.4484 49.9003 24.9158C50.7104 25.3516 51.2725 26.4424 51.2725 27.5009C51.2725 30.8645 47.7169 33.2316 42.8507 33.2316C39.233 33.2316 36.8313 31.7367 36.8313 29.5254C36.8313 28.809 36.8935 27.563 39.5762 26.3487C38.3601 25.5075 37.923 25.0406 37.923 24.2615C37.923 23.2651 38.3595 22.5487 40.9172 21.1779C39.7322 20.5552 39.3268 20.2746 38.7027 19.6209C37.923 18.8424 37.4866 17.7833 37.4866 16.6931C37.4866 13.9212 40.2936 11.8036 43.9747 11.8036C44.6599 11.8036 45.5968 11.8967 46.7818 12.1146C47.7796 12.3015 48.4971 12.3952 48.9025 12.3952C49.7132 12.3952 50.8677 12.1773 52.3331 11.7409L52.5513 13.8901L50.3984 14.326ZM45.4085 27.1898L40.5118 26.7851C40.137 27.5015 40.0754 27.7815 40.0754 28.4048C40.0754 30.4913 41.4792 31.6752 43.9742 31.6752C46.1881 31.6752 47.9668 30.4919 47.9668 28.9033C47.9662 27.7188 47.2792 27.314 45.4085 27.1898ZM40.9166 16.8806C40.9166 15.0113 42.1644 13.5164 44.0979 13.5164C46.1259 13.5164 47.3109 14.8872 47.3115 16.8806C47.3115 18.9051 46.001 20.1821 44.0357 20.1821C42.04 20.1821 40.9166 18.874 40.9166 16.8806Z"
+                    fill="currentColor"
+                  ></path>
+                  <path
+                    fillRule="evenodd"
+                    clipRule="evenodd"
+                    d="M53.2031 20.0573C53.2031 24.6048 55.4488 27.003 59.7846 27.003C64.5258 27.003 67.6132 23.8878 67.6126 18.5934C67.6126 14.1081 65.3359 11.6478 61.1878 11.6478C56.3223 11.6478 53.2031 14.9182 53.2031 20.0573ZM57.289 19.0609C57.289 15.6346 58.5063 13.2985 60.3149 13.2985C62.249 13.2985 63.5273 15.8525 63.5267 19.7146C63.5267 23.1719 62.4045 25.3833 60.4703 25.3833C58.4124 25.3833 57.289 23.0161 57.289 19.0609Z"
+                    fill="currentColor"
+                  ></path>
+                  <path
+                    d="M79.6513 15.1672L76.0019 24.8537C75.4716 26.286 74.6919 27.003 73.6313 27.003H73.1948L68.8907 14.7624C68.7036 14.2018 68.5475 13.797 68.4226 13.4854C68.2976 13.1737 68.2044 13.0495 67.8301 12.831V12.1767H73.4137V12.831C72.6334 12.8931 72.353 13.2048 72.353 13.8591C72.353 14.1701 72.4152 14.5755 72.5713 14.9492L75.129 22.2687L77.0942 15.6036C77.3118 14.7624 77.4368 14.1707 77.4368 13.797C77.4368 13.2675 77.0942 12.9558 76.2207 12.8316V12.1773H81.3666V12.8316C80.4316 13.3916 80.2127 13.7343 79.6513 15.1672Z"
+                    fill="currentColor"
+                  ></path>
+                  <path
+                    d="M83.6106 26.7845C82.5195 26.7845 81.646 25.8812 81.646 24.76C81.646 23.6078 82.5811 22.7045 83.7667 22.7045C84.8584 22.7045 85.7319 23.6078 85.7319 24.729C85.7319 25.8501 84.7956 26.7845 83.6106 26.7845Z"
+                    fill="currentColor"
+                  ></path>
+                  <path
+                    fillRule="evenodd"
+                    clipRule="evenodd"
+                    d="M91.0331 6.38388C91.0331 7.38089 91.7817 8.19044 92.7485 8.19044C93.6841 8.19044 94.4638 7.38029 94.4638 6.38388C94.4638 5.35582 93.6835 4.54626 92.7485 4.54626C91.7817 4.54626 91.0331 5.35582 91.0331 6.38388ZM89.0996 25.6012V26.3487H95.9609V25.6012C94.8381 25.3833 94.2456 25.0716 94.0273 24.511C93.9024 24.169 93.8091 23.3898 93.8091 22.2687V16.2263L94.059 12.1773L93.6531 11.7415L89.9414 12.6137V13.2364C91.1581 13.7654 91.4702 14.4507 91.4702 16.6C91.4702 16.7552 91.4702 16.9421 91.4385 17.0979L91.252 22.2681C91.127 24.8227 90.8149 25.3206 89.0996 25.6012Z"
+                    fill="currentColor"
+                  ></path>
+                  <path
+                    fillRule="evenodd"
+                    clipRule="evenodd"
+                    d="M97.8933 19.7457C97.8933 23.8567 100.139 26.6913 103.383 26.6913C106.128 26.6913 109.029 25.2275 109.404 23.6693L108.779 23.5451C106.752 24.2925 106.035 24.4794 104.881 24.4794C101.979 24.4794 100.264 22.3928 100.264 18.8734V18.0639H109.092C109.247 17.1606 109.309 16.3504 109.309 15.7278C109.309 13.2675 107.688 11.8346 104.85 11.8346C100.575 11.8346 97.8933 14.8561 97.8933 19.7457ZM104.288 13.2358C105.785 13.2358 106.939 14.2949 106.939 15.603H106.939V16.2567L106.502 16.7863H100.483C101.2 14.4818 102.541 13.2358 104.288 13.2358Z"
+                    fill="currentColor"
+                  ></path>
+                  <g clipPath="url(#clip0_1489_709)">
+                    <path
+                      d="M7.26326 9.08703C7.28351 9.09145 7.30454 9.093 7.32453 9.093C7.49197 9.093 7.63371 8.97151 7.66331 8.80588C7.6968 8.61767 7.57141 8.43725 7.38449 8.40272C7.19031 8.36819 7.0169 8.49851 6.98367 8.68543C6.95096 8.87286 7.07609 9.05432 7.26326 9.08703ZM7.32297 8.58055C7.33257 8.58055 7.34218 8.58133 7.35075 8.58288C7.43901 8.59898 7.49768 8.68491 7.48211 8.77291C7.46679 8.86118 7.38839 8.92115 7.29597 8.90635C7.20771 8.89077 7.14955 8.80536 7.16513 8.71684C7.17863 8.6374 7.24509 8.58055 7.32297 8.58055ZM16.4729 10.6742C16.4934 10.6779 16.5129 10.6797 16.5332 10.6797C16.7003 10.6797 16.8429 10.5585 16.873 10.3921C16.9054 10.2033 16.7795 10.0226 16.5921 9.98967C16.4005 9.9541 16.225 10.0855 16.192 10.2721C16.1591 10.4606 16.2852 10.6413 16.4729 10.6742ZM16.5324 10.1683C16.5409 10.1683 16.5505 10.1685 16.5596 10.1704C16.6484 10.1864 16.7071 10.2719 16.691 10.3604C16.6759 10.4486 16.5921 10.5104 16.5054 10.4941C16.4166 10.478 16.3582 10.3928 16.374 10.304C16.3867 10.2246 16.454 10.1683 16.5324 10.1683ZM13.5288 11.303C13.549 11.3066 13.5685 11.309 13.589 11.309C13.7567 11.309 13.9003 11.1875 13.9293 11.0211C13.9623 10.8323 13.8364 10.6522 13.6492 10.6187C13.4558 10.5834 13.2803 10.7137 13.2484 10.9014C13.2167 11.0896 13.3426 11.27 13.5288 11.303ZM13.5885 10.7965C13.5981 10.7965 13.6072 10.797 13.6168 10.7983C13.705 10.8144 13.7629 10.8996 13.7471 10.9886C13.7344 11.0673 13.6671 11.1239 13.589 11.1239C13.5794 11.1239 13.5703 11.1236 13.561 11.1218C13.473 11.1057 13.4143 11.0198 13.4299 10.9312C13.4428 10.8534 13.5103 10.7965 13.5885 10.7965ZM15.0212 11.052C15.0423 11.0553 15.062 11.0574 15.0817 11.0574C15.2507 11.0574 15.393 10.9362 15.421 10.77C15.4376 10.6787 15.4169 10.5855 15.3636 10.5099C15.3115 10.4339 15.2325 10.3832 15.1414 10.3674C14.9522 10.3282 14.7725 10.4619 14.7416 10.6488C14.7089 10.837 14.8338 11.0174 15.0212 11.052ZM15.0804 10.5444C15.0898 10.5444 15.0994 10.5457 15.1092 10.547C15.1518 10.5546 15.1887 10.5779 15.2128 10.6135C15.2383 10.6493 15.2476 10.6935 15.2398 10.7378C15.2248 10.8256 15.136 10.8837 15.0542 10.8708C14.9659 10.8549 14.907 10.7687 14.9234 10.6797C14.9358 10.6018 15.0025 10.5444 15.0804 10.5444ZM17.4363 11.0675C17.4031 11.2573 17.5282 11.4372 17.7151 11.4707C17.7354 11.4746 17.7559 11.4767 17.7759 11.4767C17.9436 11.4767 18.0866 11.3552 18.1165 11.189C18.1489 10.9998 18.0228 10.8196 17.8361 10.7864C17.6437 10.7524 17.4688 10.8819 17.4363 11.0675ZM17.7761 10.9647C17.785 10.9647 17.7943 10.9653 17.8039 10.9663C17.8917 10.9824 17.9506 11.0675 17.9355 11.1563C17.9215 11.235 17.8543 11.2916 17.7761 11.2916C17.7668 11.2916 17.7574 11.2911 17.7484 11.2892C17.6603 11.2731 17.6014 11.188 17.6175 11.0992C17.6313 11.0211 17.698 10.9647 17.7761 10.9647ZM9.59839 9.25214C9.61838 9.25525 9.63785 9.25681 9.65784 9.25681C9.82502 9.25681 9.9678 9.13558 9.99817 8.96995C10.0306 8.7807 9.90576 8.6008 9.71858 8.56731C9.53167 8.5333 9.35073 8.66206 9.31724 8.84975C9.28531 9.03745 9.41095 9.21865 9.59839 9.25214ZM9.65732 8.74487C9.66666 8.74487 9.67601 8.74617 9.68613 8.74721C9.77336 8.76331 9.83229 8.84846 9.81697 8.93724C9.80114 9.02551 9.71391 9.0847 9.63032 9.07068C9.54205 9.05484 9.4839 8.96969 9.49922 8.88065C9.51298 8.80173 9.57944 8.74487 9.65732 8.74487ZM10.8068 10.2189C10.8255 10.222 10.8442 10.2233 10.8624 10.2233C11.0179 10.2233 11.1506 10.1109 11.1786 9.95644C11.209 9.78095 11.0924 9.61351 10.9185 9.58235C10.7451 9.55042 10.5766 9.67062 10.5459 9.84481C10.5158 10.0203 10.6324 10.188 10.8068 10.2189ZM10.8621 9.74798C10.8702 9.74798 10.8793 9.74902 10.8884 9.75032C10.9693 9.76537 11.0249 9.84455 11.0098 9.92762C10.9956 10.0091 10.9143 10.0644 10.8367 10.0517C10.7549 10.0372 10.7001 9.95774 10.7149 9.87519C10.7271 9.80068 10.7892 9.74798 10.8621 9.74798ZM8.554 8.11742C8.57451 8.12131 8.59398 8.12339 8.61475 8.12339C8.78168 8.12339 8.92394 8.00215 8.95405 7.83549C8.98702 7.64676 8.86086 7.46633 8.67342 7.43336C8.48184 7.3978 8.3066 7.52864 8.27337 7.71477C8.24014 7.90376 8.36631 8.08471 8.554 8.11742ZM8.61294 7.61145C8.62202 7.61145 8.63163 7.61171 8.64123 7.61379C8.7295 7.62988 8.78869 7.71503 8.77233 7.80382C8.75857 7.88248 8.69185 7.93907 8.61501 7.93907C8.60541 7.93907 8.59606 7.93829 8.58671 7.93647C8.49845 7.92064 8.43926 7.83497 8.45536 7.74592C8.46808 7.66778 8.53505 7.61145 8.61294 7.61145ZM4.68904 31.4269C5.24199 31.4269 5.71343 31.0268 5.80923 30.477C5.918 29.8519 5.50367 29.2548 4.88582 29.1452C4.8191 29.1338 4.75212 29.1278 4.68567 29.1278C4.13297 29.1278 3.66205 29.5271 3.56626 30.0777C3.45723 30.7028 3.87103 31.2999 4.48915 31.4103C4.55664 31.4206 4.62362 31.4269 4.68904 31.4269ZM3.77394 30.1133C3.85182 29.6644 4.23603 29.3376 4.68593 29.3376C4.73966 29.3376 4.79444 29.3423 4.8487 29.3521C5.35284 29.442 5.69085 29.9295 5.60206 30.4396C5.51432 30.943 5.0408 31.2945 4.52653 31.2018C4.02264 31.112 3.68594 30.6229 3.77394 30.1133ZM4.68956 31.0572C5.0652 31.0572 5.38426 30.7849 5.44968 30.4116C5.5234 29.9871 5.24199 29.5819 4.82248 29.5082C4.3918 29.4305 3.99954 29.7226 3.92659 30.1416C3.85312 30.566 4.13453 30.9713 4.55301 31.0445C4.5987 31.0525 4.64491 31.0572 4.68956 31.0572ZM4.10753 30.1733C4.15711 29.8872 4.4001 29.6798 4.68618 29.6798C4.72045 29.6798 4.75524 29.6829 4.79029 29.6904C5.11089 29.746 5.32507 30.0562 5.26873 30.3809C5.21292 30.701 4.91048 30.9271 4.5852 30.8638C4.26537 30.8077 4.05094 30.498 4.10753 30.1733ZM5.776 10.4663C5.35129 10.3887 4.95358 10.6802 4.88063 11.1C4.80716 11.5237 5.08831 11.9297 5.50731 12.0042C5.55248 12.0125 5.59791 12.0164 5.64256 12.0164C6.01821 12.0164 6.33882 11.7448 6.40398 11.3702C6.4777 10.9458 6.19629 10.54 5.776 10.4663ZM6.22277 11.3391C6.16748 11.6584 5.8653 11.8785 5.53976 11.8232C5.22045 11.7659 5.00524 11.4551 5.06209 11.1306C5.11115 10.8445 5.35466 10.6379 5.641 10.6379C5.67501 10.6379 5.71006 10.6405 5.74381 10.6465C6.06467 10.7036 6.27937 11.0143 6.22277 11.3391ZM12.0664 27.7229C12.2318 27.7229 12.3655 27.5876 12.3655 27.4225C12.3655 27.2558 12.2315 27.1214 12.0664 27.1214C11.9013 27.1214 11.7676 27.2561 11.7676 27.4225C11.7674 27.5881 11.9013 27.7229 12.0664 27.7229ZM12.0664 27.3008C12.1329 27.3008 12.1866 27.3555 12.1866 27.4225C12.1866 27.4887 12.1329 27.5435 12.0664 27.5435C12.0002 27.5435 11.947 27.4887 11.947 27.4225C11.947 27.354 11.9989 27.3008 12.0664 27.3008ZM11.0483 26.706C10.8837 26.706 10.7497 26.8413 10.7497 27.0071C10.7497 27.173 10.8839 27.3075 11.0483 27.3075C11.2139 27.3075 11.3478 27.1728 11.3478 27.0071C11.3478 26.8415 11.2139 26.706 11.0483 26.706ZM11.0483 27.1289C10.9826 27.1289 10.9291 27.0749 10.9291 27.0071C10.9291 26.9402 10.9826 26.8854 11.0483 26.8854C11.115 26.8854 11.1682 26.9402 11.1682 27.0071C11.1682 27.0749 11.115 27.1289 11.0483 27.1289ZM13.0688 27.5199C12.9036 27.5199 12.7697 27.6551 12.7697 27.821C12.7697 27.9877 12.9039 28.1221 13.0688 28.1221C13.2336 28.1221 13.3678 27.9869 13.3678 27.821C13.3678 27.6551 13.2339 27.5199 13.0688 27.5199ZM13.0688 27.9428C13.0018 27.9428 12.9491 27.8888 12.9491 27.821C12.9491 27.754 13.0026 27.6985 13.0688 27.6985C13.135 27.6985 13.1884 27.7532 13.1884 27.821C13.1884 27.888 13.1342 27.9428 13.0688 27.9428ZM16.4867 18.7276C16.3299 18.7276 16.2027 18.8564 16.2027 19.0142C16.2027 19.1723 16.3299 19.3003 16.4867 19.3003C16.644 19.3003 16.772 19.1723 16.772 19.0142C16.772 18.8564 16.6443 18.7276 16.4867 18.7276ZM16.4867 19.1173C16.4314 19.1173 16.3862 19.0714 16.3862 19.0142C16.3862 18.9574 16.4314 18.9114 16.4867 18.9114C16.543 18.9114 16.5882 18.9579 16.5882 19.0142C16.5882 19.0714 16.543 19.1173 16.4867 19.1173ZM15.5643 18.3595C15.4067 18.3595 15.2788 18.4883 15.2788 18.6459C15.2788 18.804 15.4065 18.9325 15.5643 18.9325C15.7211 18.9325 15.8486 18.8037 15.8486 18.6459C15.8486 18.4883 15.7211 18.3595 15.5643 18.3595ZM15.5643 18.7494C15.5085 18.7494 15.4623 18.7027 15.4623 18.6459C15.4623 18.589 15.5085 18.5428 15.5643 18.5428C15.6196 18.5428 15.6648 18.5898 15.6648 18.6459C15.6648 18.7027 15.6196 18.7494 15.5643 18.7494ZM14.6264 17.9776C14.4691 17.9776 14.3411 18.1056 14.3411 18.2637C14.3411 18.4213 14.4688 18.5501 14.6264 18.5501C14.7834 18.5501 14.9114 18.421 14.9114 18.2637C14.9114 18.1056 14.7837 17.9776 14.6264 17.9776ZM14.6264 18.367C14.5706 18.367 14.5243 18.3201 14.5243 18.2637C14.5243 18.2061 14.5706 18.1607 14.6264 18.1607C14.6822 18.1607 14.7276 18.2064 14.7276 18.2637C14.7276 18.3201 14.6822 18.367 14.6264 18.367ZM1.65221 24.7003C1.7345 24.8106 1.7127 24.9687 1.60237 25.0523C1.55823 25.0858 1.50631 25.1032 1.45231 25.1032C1.44063 25.1032 1.42869 25.1017 1.41675 25.1009C1.35081 25.091 1.29188 25.0567 1.2519 25.0028C1.16935 24.8916 1.19115 24.7343 1.30071 24.6512C1.35393 24.6102 1.41935 24.5926 1.48554 24.6019C1.55304 24.6108 1.61119 24.6455 1.65221 24.7003ZM7.41695 31.9591C7.39618 31.931 7.36087 31.9149 7.3266 31.9188C7.29104 31.9227 7.26066 31.944 7.24639 31.9759L6.0587 34.5481C6.04027 34.5883 6.05118 34.6371 6.08492 34.6652C6.10362 34.6813 6.12568 34.6888 6.14879 34.6888C6.16852 34.6888 6.18876 34.6828 6.20668 34.6706C6.4385 34.5055 6.82687 34.1678 7.03922 33.9481C7.30194 33.6743 7.76274 33.0637 7.94835 32.8134C7.97405 32.7784 7.97405 32.7314 7.94887 32.6963L7.41695 31.9591ZM6.89618 33.8103C6.77339 33.938 6.58881 34.107 6.41306 34.2552L7.35568 32.214L7.74664 32.754C7.54155 33.0271 7.13268 33.5647 6.89618 33.8103ZM18.8322 12.3804C19.1136 11.3684 18.8179 10.6844 18.5194 10.2874C18.2987 9.99538 18.0191 9.78666 17.8595 9.71034V9.7093C16.7315 8.94243 15.353 9.45723 15.312 9.47618C14.5833 9.93697 13.8187 10.0683 13.0376 9.86818C10.9366 9.33132 9.28271 6.63794 8.97196 6.10264L8.79751 5.89262C8.74793 5.83655 8.68121 5.79994 8.60541 5.7854L8.19991 5.52139C8.09866 5.45259 7.96471 5.45675 7.86943 5.52918L7.82816 5.56007C7.82478 5.55929 7.82141 5.55799 7.81725 5.55773C7.62463 5.51646 7.34919 5.592 7.08413 5.76334C6.81934 5.93831 6.56129 6.21479 6.59037 6.48581L6.05844 6.8843C5.87854 6.88975 5.66982 6.9653 5.46136 7.10315C5.25004 7.24022 5.02782 7.45984 4.97798 7.69297L4.63271 7.95205L4.57637 8.00163L4.57014 8.03227C4.55119 8.06238 4.53899 8.09327 4.53276 8.12572C4.5216 8.15506 4.52056 8.17972 4.52212 8.18959L4.52627 8.27551L4.55119 8.30615C4.59065 8.38507 4.85025 8.90453 5.02133 9.22774C4.61168 9.59585 4.03069 10.1753 3.40245 10.9785C3.07431 10.9793 2.79109 11.2049 2.7184 11.5369C2.69114 11.6607 2.69659 11.7846 2.73553 11.8993C1.694 13.4546 0.941157 15.1859 0.498015 17.049C0.498015 17.049 0.477766 17.1233 0.459075 17.2212C0.183377 18.4379 0.0424126 19.7123 0.0413742 21.0108C0.029692 21.0368 0.0465662 21.5874 0.0527967 21.738C0.0468258 21.8634 -0.0855715 24.8524 1.09458 28.0385C1.08731 28.1185 1.09458 28.2021 1.11613 28.2888C1.14183 28.3929 1.20725 28.6593 1.40896 28.8228C1.86223 29.8594 2.42505 30.8072 3.08755 31.6468C3.57768 32.2072 4.23448 33.2508 4.10078 33.8729C3.93749 33.8165 3.77368 33.8015 3.61922 33.8295C3.32924 33.8835 3.09378 34.0759 2.91855 34.4006C2.7659 34.6867 2.75111 35.2729 3.12026 35.6296C3.39492 35.8965 3.80561 35.9582 4.30431 35.7931C4.30846 35.803 4.31288 35.8136 4.31833 35.8227C4.37129 35.9193 4.46864 35.9855 4.58001 35.9993C4.65425 36.0076 4.73681 36.013 4.82351 36.013C5.7207 36.013 7.14566 35.479 8.86475 32.9533C9.46495 33.0613 9.98493 33.1426 10.4058 33.193L11.0857 33.3116C11.3702 33.4043 11.9891 33.3695 12.3068 32.5639L19.1614 14.8328C19.2478 14.6493 19.3093 14.453 19.3452 14.2505C19.4594 13.5865 19.2696 12.8977 18.8322 12.3804ZM8.54051 6.36691C8.61605 6.50321 10.4226 9.7241 12.9114 10.3609C13.8369 10.5982 14.7344 10.4445 15.5513 9.92269C15.5628 9.9175 16.6951 9.53407 17.5967 10.1434C17.6445 10.1654 18.7294 10.7088 18.3571 12.181C18.3174 12.1506 18.2808 12.1156 18.2382 12.0883C18.3605 11.7513 18.2382 11.4811 18.2328 11.4694C18.2211 11.4445 18.1982 11.4263 18.1715 11.4203C18.1445 11.4128 18.1167 11.419 18.0941 11.4362C18.0142 11.4974 17.9199 11.5299 17.821 11.5299C17.5723 11.5299 17.3701 11.3245 17.3701 11.072C17.3701 10.946 17.4231 10.823 17.516 10.7352C17.536 10.7166 17.5459 10.6906 17.5448 10.6631C17.5433 10.6361 17.5295 10.6106 17.5069 10.5945L17.1152 10.3059C17.09 10.2877 17.0578 10.2815 17.0293 10.2942C16.9997 10.3048 16.9789 10.3297 16.9714 10.3601C16.9241 10.5642 16.7401 10.7132 16.5326 10.7132C16.3018 10.7132 16.109 10.5382 16.083 10.3072C16.0801 10.2771 16.0628 10.2511 16.0368 10.2366C16.0103 10.222 15.9794 10.2207 15.9524 10.2342L15.5596 10.4196C15.5202 10.4377 15.4994 10.4814 15.509 10.5224C15.5173 10.56 15.5217 10.5974 15.5217 10.6324C15.5217 10.8845 15.3185 11.0899 15.0706 11.0899C14.8618 11.0899 14.6806 10.9458 14.631 10.7391C14.6251 10.714 14.6098 10.6932 14.5874 10.6805C14.5667 10.6688 14.5394 10.6646 14.5153 10.6729L14.1098 10.7973C14.0711 10.8085 14.0451 10.8451 14.0451 10.885C14.0451 11.1363 13.8426 11.3414 13.5937 11.3414C13.3447 11.3414 13.1425 11.1363 13.1425 10.885C13.1425 10.8393 13.149 10.7955 13.1607 10.7547C13.1677 10.7298 13.1645 10.7031 13.1508 10.6807C13.137 10.6589 13.1155 10.6433 13.0898 10.6387L12.7393 10.5727C12.7123 10.5673 12.684 10.5743 12.6638 10.5919C12.6425 10.6096 12.6303 10.6356 12.6303 10.6623C12.6303 10.9788 12.366 11.0263 12.2092 11.0263C12.2004 11.0263 12.1926 11.0211 12.1832 11.0208C12.3398 11.0099 12.4714 10.8962 12.5 10.7373C12.5324 10.5481 12.4062 10.3677 12.2196 10.3347C12.0327 10.3017 11.8533 10.4294 11.818 10.6171C11.8029 10.7059 11.8258 10.7921 11.8702 10.863C11.8027 10.783 11.7572 10.6831 11.7572 10.5701C11.7572 10.5039 11.7713 10.438 11.7983 10.3788C11.8089 10.3557 11.8089 10.3298 11.799 10.3053C11.7892 10.2815 11.7694 10.2633 11.7458 10.2547L11.3416 10.1083C11.3011 10.093 11.2565 10.1096 11.2336 10.1449C11.15 10.2742 11.0091 10.3516 10.8564 10.3516C10.8432 10.3516 10.832 10.3453 10.8198 10.344L10.5799 10.2467C10.4761 10.1628 10.4045 10.039 10.4045 9.89466C10.4045 9.81184 10.4247 9.73448 10.4657 9.66439C10.4901 9.62311 10.479 9.57119 10.4408 9.54315L10.0397 9.25499C10.0057 9.23059 9.95949 9.23137 9.92704 9.25889C9.84735 9.32587 9.74714 9.36169 9.64304 9.36169C9.39408 9.36169 9.19185 9.15686 9.19185 8.90479C9.19185 8.7781 9.24091 8.66102 9.33048 8.57509C9.36474 8.54264 9.36838 8.48969 9.33956 8.45282L9.05815 8.08886C9.04206 8.06835 9.0174 8.05485 8.99118 8.05381C8.96599 8.05148 8.93874 8.06238 8.92004 8.08159C8.83541 8.167 8.72223 8.21503 8.60099 8.21503C8.35255 8.21503 8.15032 8.0102 8.1498 7.75916C8.1498 7.74307 8.14617 7.72723 8.13786 7.71218V7.71192H8.1376V7.71088L8.13656 7.71062C8.13449 7.7062 8.13163 7.70309 8.12903 7.69868L8.12852 7.69816C8.11969 7.68751 8.10879 7.67947 8.09711 7.67324H8.09684C8.0862 7.66856 8.0753 7.66571 8.0631 7.66519H8.06258C8.04804 7.66571 8.03142 7.66778 8.01637 7.67531H8.01611H8.01533V7.67557H8.01429H8.01377C8.00832 7.67869 8.00365 7.68232 7.99897 7.68596L7.41824 8.1535C7.39566 8.17115 7.38294 8.19893 7.38398 8.22775C7.38501 8.25682 7.39929 8.28382 7.42266 8.29992C7.42577 8.30199 7.7103 8.50734 7.72509 8.77836C7.73314 8.93023 7.655 9.08028 7.49197 9.22488C7.45926 9.25448 7.45173 9.30302 7.47432 9.3404C7.47639 9.3443 7.47951 9.34845 7.48211 9.35209L7.4982 9.37285C8.07763 10.1252 8.14721 10.1595 8.1807 10.1763C8.19939 10.1857 8.22068 10.1877 8.24144 10.1841C8.27182 10.1815 8.31803 10.1597 8.33075 10.1218C8.39046 9.94138 8.55738 9.82015 8.74507 9.82015C8.75624 9.82015 8.7648 9.82534 8.77648 9.82638C8.6093 9.82638 8.46314 9.94112 8.43485 10.1062C8.4024 10.2955 8.52727 10.4767 8.7147 10.5094C8.73495 10.5128 8.75468 10.5154 8.77441 10.5154C8.94315 10.5154 9.08619 10.3939 9.11475 10.2277C9.13577 10.1078 9.08749 9.9946 9.00519 9.91698C9.10955 9.99798 9.18146 10.1192 9.18146 10.2623C9.18146 10.3817 9.13266 10.4985 9.04725 10.5803C9.027 10.6 9.01714 10.6286 9.02025 10.6563C9.02311 10.6844 9.03868 10.709 9.06283 10.7243L9.63343 11.0792C9.65939 11.0953 9.69288 11.0966 9.72274 11.0839C9.75051 11.0696 9.77024 11.0429 9.77388 11.0117C9.78997 10.87 9.87642 10.756 9.99091 10.688C9.92808 10.7397 9.88083 10.8111 9.86604 10.8965C9.83255 11.0852 9.95794 11.2656 10.1451 11.2991C10.1656 11.3033 10.1856 11.3051 10.2064 11.3051C10.3733 11.3051 10.5161 11.1836 10.5462 11.0172C10.5701 10.8809 10.5086 10.7506 10.4032 10.675C10.5446 10.7475 10.6441 10.8926 10.6441 11.0634C10.6441 11.1602 10.6121 11.2529 10.5527 11.3336C10.535 11.3573 10.5298 11.3884 10.5384 11.4164C10.548 11.4453 10.5701 11.4673 10.5989 11.4759L11.2227 11.6696C11.25 11.6794 11.2804 11.6727 11.3042 11.6558C11.3281 11.6387 11.3414 11.6111 11.3414 11.5818C11.3414 11.3806 11.4771 11.2178 11.6599 11.1646C11.5662 11.2129 11.493 11.2991 11.4727 11.4105C11.4397 11.5989 11.5659 11.7796 11.7531 11.8126C11.7733 11.8162 11.7936 11.8181 11.8133 11.8181C11.981 11.8181 12.1241 11.6973 12.1537 11.5307C12.1672 11.4512 12.1477 11.3759 12.1132 11.3095C12.1734 11.385 12.2147 11.4777 12.2147 11.5818C12.2147 11.6613 12.1944 11.7376 12.1552 11.804C12.1407 11.8277 12.1378 11.8565 12.1479 11.8824C12.1575 11.9084 12.1773 11.9289 12.2043 11.9382C12.2149 11.9427 12.3463 11.9837 12.5192 11.9912C12.5033 12.0182 12.4914 12.0473 12.4854 12.0795C12.4854 12.0803 12.4854 12.0813 12.4854 12.0813C12.4109 12.0839 12.3356 12.0844 12.258 12.0839C12.2398 11.978 12.1599 11.8892 12.048 11.8684C11.9107 11.843 11.7902 11.9312 11.7552 12.0579C11.6884 12.0522 11.6212 12.0457 11.5532 12.0372C11.5462 12.0003 11.5332 11.9652 11.5106 11.9336C11.4712 11.877 11.4115 11.8391 11.3437 11.8274C11.229 11.804 11.123 11.8645 11.0709 11.9598C10.9782 11.9406 10.8821 11.917 10.784 11.8902C10.7856 11.8855 10.7881 11.8817 10.7894 11.8759C10.8141 11.7326 10.7191 11.5961 10.5773 11.5712C10.4546 11.5514 10.3401 11.6231 10.2939 11.7342C10.2264 11.7095 10.1599 11.6838 10.094 11.6566C10.0971 11.6467 10.1025 11.6376 10.1046 11.6275C10.1285 11.4855 10.0335 11.3487 9.89226 11.3232C9.784 11.3051 9.68068 11.3567 9.6259 11.4453C9.5218 11.3944 9.41615 11.3388 9.30633 11.2781C9.30945 11.2687 9.31438 11.2609 9.31646 11.2506C9.34138 11.1083 9.24688 10.971 9.10462 10.946C9.01636 10.93 8.93536 10.9645 8.87695 11.0237C8.79128 10.9694 8.70899 10.9152 8.62747 10.8599C8.63864 10.8375 8.64928 10.8147 8.65421 10.789C8.67706 10.6459 8.58152 10.5099 8.44108 10.4847C8.35333 10.4681 8.27182 10.5029 8.21393 10.5624C8.09762 10.4741 7.98677 10.3874 7.88397 10.3033C7.89773 10.2786 7.90837 10.2521 7.91357 10.2236C7.92577 10.1535 7.91045 10.0839 7.86969 10.0258C7.82919 9.96786 7.76897 9.92944 7.70095 9.91802C7.62177 9.9014 7.54986 9.92425 7.49457 9.9689C7.32712 9.81833 7.19888 9.69424 7.10438 9.59871C7.12048 9.57171 7.13242 9.54186 7.13787 9.50863C7.14981 9.43853 7.1345 9.3687 7.09374 9.31055C7.05402 9.25318 6.99405 9.21579 6.92604 9.20307C6.77884 9.17815 6.64826 9.27602 6.6236 9.41699C6.60854 9.50058 6.63892 9.57924 6.69239 9.63765C6.68487 9.65348 6.67422 9.6675 6.67241 9.68594C6.21291 9.19113 5.90346 8.76798 5.76717 8.5699L8.52493 6.34563L8.54051 6.36691ZM15.1355 11.5657C15.138 11.5574 15.1425 11.5499 15.1443 11.5421C15.1521 11.4972 15.1445 11.4531 15.1303 11.4113C15.6435 11.1652 16.3011 10.9837 16.6793 11.189C16.8564 11.2848 16.9654 11.4673 17.009 11.7243C16.9563 11.7189 16.9075 11.7173 16.8571 11.7132C16.8047 11.4393 16.6627 11.3035 16.5394 11.2454C16.1557 11.0613 15.549 11.3295 15.1355 11.5657ZM16.5241 11.6994C15.8377 11.7025 15.3509 11.8887 15.0802 12.2648L15.0765 11.9829C15.5165 11.7033 16.1269 11.4126 16.3992 11.5392C16.4239 11.5525 16.4862 11.5844 16.5241 11.6994ZM16.7585 11.0403C16.4075 10.8492 15.9163 10.9398 15.5285 11.0681C15.6349 10.9533 15.7035 10.8014 15.7035 10.6319C15.7035 10.6065 15.7014 10.5808 15.6983 10.5548L15.9239 10.4484C16.0054 10.7111 16.2478 10.8957 16.5303 10.8957C16.7777 10.8957 17.0012 10.7467 17.1048 10.5273L17.3161 10.6828C17.232 10.795 17.1853 10.9318 17.1853 11.0717C17.1853 11.425 17.4701 11.7132 17.82 11.7132C17.9179 11.7132 18.0113 11.6914 18.0967 11.6488C18.1144 11.7306 18.1165 11.8583 18.0705 12.0006C17.888 11.9055 17.6941 11.8271 17.4815 11.7898C17.3782 11.7716 17.2821 11.7594 17.1842 11.7461C17.1388 11.4079 16.9971 11.169 16.7585 11.0403ZM16.4906 16.7476C16.8595 16.7476 17.1645 16.4577 17.2068 16.0844C17.2733 16.0812 17.3732 16.0758 17.4776 16.0636L15.7803 18.0963C15.7606 18.0888 15.7406 18.0812 15.7198 18.0755L15.8013 16.2425C15.8961 16.535 16.1692 16.7476 16.4906 16.7476ZM15.8242 15.7204L15.8392 15.3837C15.879 15.4392 15.9169 15.4818 15.9475 15.514C15.8961 15.5748 15.8548 15.6454 15.8242 15.7204ZM17.1175 15.9039C17.0679 15.905 17.0472 15.9398 17.0472 15.9899C17.0472 16.298 16.8112 16.5631 16.4903 16.5633C16.192 16.5646 15.9488 16.3172 15.9488 16.0158C15.9488 15.8502 16.0199 15.6615 16.1375 15.5587C16.158 15.541 16.1702 15.5374 16.1697 15.5104C16.1692 15.4831 16.1565 15.4577 16.1357 15.4403C16.1336 15.439 15.9698 15.303 15.8553 15.0457C15.8548 15.0447 15.8543 15.0436 15.8538 15.0426L15.8584 14.9341C15.9659 15.1654 16.1785 15.3824 16.3691 15.527C16.3164 15.5434 16.2647 15.5672 16.219 15.6028C16.1191 15.6809 16.0555 15.7939 16.0409 15.9198C16.0085 16.1796 16.192 16.4169 16.449 16.4507C16.449 16.4507 16.489 16.4543 16.5085 16.4543C16.6118 16.4543 16.7125 16.42 16.7959 16.3549C16.8963 16.2767 16.9599 16.1638 16.975 16.0369C16.9859 15.9476 16.9644 15.865 16.9298 15.7884C17.0285 15.8092 17.164 15.8458 17.2746 15.8458H17.2748C17.4246 15.8458 17.5716 15.8219 17.7115 15.7798C17.7115 15.7798 17.6235 15.8541 17.611 15.8564C17.394 15.8977 17.1201 15.9039 17.1175 15.9039ZM16.8574 14.8437C16.8979 14.8437 16.9371 14.8505 16.9742 14.8642C16.9906 14.8725 17.0077 14.8798 17.0246 14.8871C17.1751 14.9709 17.2426 15.156 17.1816 15.3193C17.117 15.4948 16.9093 15.5922 16.7364 15.5244C16.556 15.4559 16.4657 15.2523 16.5332 15.0698C16.5846 14.9343 16.7154 14.8437 16.8574 14.8437ZM16.8252 14.0815C16.833 14.0662 16.8403 14.0504 16.8462 14.0338C16.9542 13.8224 17.2042 13.7108 17.4472 13.8009C17.4654 13.81 17.4841 13.8191 17.5035 13.8261C17.718 13.9416 17.8161 14.1994 17.731 14.4341C17.7208 14.4528 17.7128 14.472 17.705 14.4925C17.5934 14.7033 17.3535 14.8123 17.1032 14.7196C17.0861 14.7108 17.0687 14.7033 17.0505 14.6955C16.8343 14.5815 16.7356 14.3227 16.8252 14.0815ZM16.542 15.6879C16.7032 15.7085 16.8182 15.8577 16.7977 16.0215C16.7873 16.101 16.7484 16.1721 16.685 16.2212C16.623 16.2697 16.5474 16.2905 16.469 16.2811C16.3076 16.2599 16.1926 16.1103 16.2133 15.9473C16.2227 15.8679 16.2621 15.7975 16.325 15.7482C16.3878 15.6991 16.4646 15.6773 16.542 15.6879ZM17.1282 15.6555C17.2302 15.5945 17.3117 15.4997 17.3543 15.3837C17.4109 15.2321 17.394 15.0688 17.321 14.9346C17.4542 14.926 17.5796 14.8777 17.6819 14.7991C17.7198 14.9696 17.8397 15.1184 18.0134 15.1851C18.0731 15.2074 18.1357 15.2191 18.199 15.2191C18.2315 15.2191 18.2634 15.2165 18.2946 15.2105C18.0043 15.5275 17.5659 15.703 17.1282 15.6555ZM17.5783 13.6581C17.5614 13.6495 17.543 13.6412 17.5248 13.6335C17.3761 13.5491 17.3088 13.3655 17.3696 13.203C17.4353 13.0268 17.6419 12.9304 17.8156 12.9979C17.9947 13.0665 18.0861 13.27 18.0191 13.4512C17.9537 13.6246 17.7538 13.7217 17.5783 13.6581ZM17.197 13.1389C17.1406 13.29 17.157 13.4538 17.2304 13.5872C17.097 13.5958 16.9724 13.6444 16.8699 13.7235C16.8322 13.553 16.7128 13.4042 16.5394 13.3385C16.4786 13.3154 16.415 13.3043 16.3507 13.3043C16.3198 13.3043 16.2878 13.3066 16.2575 13.3126C16.5479 12.9946 16.9864 12.8209 17.4236 12.8668C17.3223 12.9281 17.2406 13.0229 17.197 13.1389ZM16.3509 13.4876C16.393 13.4876 16.4342 13.4953 16.4742 13.5107C16.6521 13.5776 16.7432 13.7788 16.6778 13.9608C16.671 13.9761 16.6637 13.9909 16.6572 14.0062C16.5773 14.1568 16.3911 14.23 16.2299 14.1688C16.05 14.1002 15.9597 13.8967 16.0272 13.7155C16.0781 13.5789 16.2079 13.4876 16.3509 13.4876ZM16.1648 14.3406C16.225 14.3635 16.2876 14.3744 16.3519 14.3744C16.4436 14.3744 16.5316 14.3513 16.609 14.3082C16.6175 14.4398 16.6656 14.5667 16.7447 14.6721C16.5724 14.7105 16.4257 14.8352 16.3623 15.0062C16.3268 15.101 16.3198 15.1994 16.3366 15.2923C16.0186 14.9948 15.8522 14.5576 15.8979 14.1158C15.9574 14.215 16.0487 14.2954 16.1648 14.3406ZM15.974 12.8798L15.4288 14.1887L15.1453 14.546L15.9773 12.7106C15.9872 12.7334 15.9968 12.7804 15.974 12.8798ZM15.698 14.4247L15.5368 18.054C15.4914 18.056 15.4467 18.0636 15.4026 18.0758L15.5511 14.7848L15.698 14.4247ZM14.4919 21.494C14.685 21.5669 14.7824 21.786 14.711 21.9823C14.6406 22.1694 14.4249 22.2767 14.2323 22.2042C14.2157 22.1959 14.1983 22.1881 14.1812 22.1814C14.0179 22.0916 13.9449 21.8917 14.0103 21.7144C14.0809 21.5259 14.3029 21.4218 14.4919 21.494ZM14.3515 20.9311C14.3626 20.8413 14.4073 20.7632 14.4771 20.7084C14.5358 20.6627 14.6056 20.6391 14.6773 20.6391C14.6908 20.6391 14.7043 20.6399 14.7178 20.6414C14.8977 20.6651 15.0262 20.8307 15.0041 21.0127C14.9927 21.1025 14.9475 21.1814 14.8777 21.237C14.8091 21.2894 14.7245 21.3154 14.6368 21.3016C14.4569 21.2795 14.3294 21.1136 14.3515 20.9311ZM7.43927 10.3282C7.44031 10.3295 7.44239 10.3305 7.44291 10.3316C7.43538 10.3503 7.42006 10.3651 7.42006 10.3869V10.4193C7.26741 10.2841 7.12437 10.1501 6.9912 10.0177V9.72513C7.09166 9.8256 7.22432 9.95436 7.40085 10.1135C7.39981 10.119 7.39566 10.1241 7.39462 10.1296C7.38242 10.1997 7.39851 10.2698 7.43927 10.3282ZM7.41876 10.8417L7.41772 11.2713C7.29649 11.3822 7.15656 11.5146 6.98938 11.6854L6.9899 10.464C7.12541 10.5899 7.26793 10.7158 7.41876 10.8417ZM7.41643 12.0346L7.38917 24.9277L6.98367 25.61L6.98938 12.4637L7.41643 12.0346ZM7.38735 25.9272L7.38034 29.4417C7.38034 29.4578 7.38242 29.4736 7.38657 29.4882L6.98107 30.3355L6.98341 26.6066L7.38735 25.9272ZM7.15864 30.7522C7.30558 30.4388 7.6049 30.2358 7.92161 30.2358C7.94394 30.2358 7.96419 30.2418 7.98599 30.2441L7.55661 31.2109C7.41772 31.0824 7.28325 30.9508 7.15397 30.8168C7.15007 30.7947 7.14956 30.7727 7.15864 30.7522ZM8.87306 23.428L8.87565 26.3296C8.87565 26.3418 8.87669 26.354 8.87903 26.3646L8.4242 27.3161L8.43614 24.1629L8.87306 23.428ZM8.44004 23.1611L8.48054 11.8492C8.58204 11.9307 8.7108 12.0146 8.86397 12.0997L8.87254 22.435L8.44004 23.1611ZM8.86293 11.4172V11.4723C8.73261 11.3967 8.60593 11.3165 8.4821 11.235L8.48313 10.9899C8.48962 10.9876 8.49534 10.9821 8.50157 10.979C8.59736 11.0447 8.69705 11.1093 8.79855 11.1737C8.78946 11.2552 8.81984 11.3313 8.87565 11.3871C8.87306 11.3975 8.86293 11.4055 8.86293 11.4172ZM8.42109 10.8274C8.39305 10.8046 8.36008 10.7859 8.32062 10.7854H8.31984C8.31829 10.7854 8.31725 10.7877 8.31465 10.7877C8.30453 10.7695 8.297 10.749 8.30089 10.7267C8.30842 10.6815 8.34632 10.6491 8.40551 10.6491C8.40681 10.6491 8.40785 10.6491 8.40941 10.6491C8.46029 10.6576 8.493 10.707 8.48599 10.7576C8.4795 10.7937 8.45224 10.8165 8.42109 10.8274ZM9.20976 26.4345L9.02518 27.9038L8.8689 28.2566L7.72847 29.5313L9.20976 26.4345ZM8.9647 11.1875C8.97223 11.1436 9.01091 11.1101 9.06906 11.1101C9.07035 11.1101 9.07191 11.1101 9.07217 11.1101C9.12228 11.1189 9.15706 11.1696 9.14772 11.221C9.14304 11.25 9.1228 11.27 9.09891 11.2835C9.07607 11.2713 9.05478 11.2534 9.02622 11.2534C9.01376 11.2534 9.00519 11.2651 8.99273 11.2682C8.97326 11.2461 8.95873 11.2189 8.9647 11.1875ZM10.9782 22.7351L10.784 23.9293L9.85254 25.09L10.9782 22.7351ZM10.4366 11.8129C10.4442 11.7679 10.4818 11.7355 10.541 11.7355C10.542 11.7355 10.5431 11.7355 10.5452 11.7355C10.5953 11.7443 10.6293 11.7939 10.6204 11.8456C10.6142 11.8843 10.5825 11.9089 10.5472 11.9175C10.5301 11.9118 10.5179 11.8967 10.4982 11.8967H10.4977C10.4917 11.8967 10.4883 11.9024 10.4823 11.9032C10.4525 11.883 10.4304 11.8515 10.4366 11.8129ZM12.2733 11.2077C12.5685 11.1875 12.7611 11.0265 12.8037 10.7716L12.9618 10.8014C12.9584 10.8284 12.9566 10.8562 12.9566 10.8845C12.9566 11.1296 13.095 11.3443 13.2972 11.451C13.2095 11.4889 13.1274 11.5473 13.0493 11.627C12.8338 11.8432 12.5202 11.8196 12.3608 11.7885C12.3837 11.7228 12.3959 11.6532 12.3959 11.5813C12.3964 11.4414 12.3502 11.3121 12.2733 11.2077ZM12.9532 23.6372C12.9138 23.7429 12.9083 23.8517 12.9299 23.9545C12.5771 23.6336 12.3935 23.1536 12.4491 22.67C12.5109 22.7819 12.6108 22.873 12.7385 22.9218C12.8016 22.9462 12.8683 22.9576 12.9348 22.9576C13.0342 22.9576 13.1303 22.9316 13.2136 22.8823C13.2217 23.0277 13.2749 23.1684 13.3642 23.2831C13.1783 23.3208 13.0202 23.454 12.9532 23.6372ZM13.0384 24.3483C13.1363 24.3605 13.2235 24.4106 13.2855 24.489C13.3481 24.5697 13.3761 24.6699 13.3631 24.7735C13.3385 24.9796 13.1503 25.1333 12.9449 25.1035C12.7396 25.0775 12.5934 24.8859 12.6181 24.6782C12.642 24.4882 12.8019 24.346 12.9904 24.346C13.0059 24.346 13.022 24.3467 13.0384 24.3483ZM12.3912 21.6757C12.3642 21.5145 12.285 21.3797 12.1599 21.3045L12.514 20.5075L12.3912 21.6757ZM13.2899 22.5176C13.2793 22.5381 13.2697 22.5594 13.2611 22.5822C13.1718 22.7395 12.9732 22.8135 12.8042 22.7489C12.6106 22.6759 12.5129 22.4561 12.5859 22.2606C12.6407 22.1131 12.7814 22.015 12.9356 22.015C12.9802 22.015 13.0254 22.0233 13.0675 22.0402C13.2564 22.11 13.3548 22.3244 13.2899 22.5176ZM12.6363 21.0963L12.7383 20.1303L13.2728 19.5311L12.6363 21.0963ZM12.9011 12.8759C13.021 12.8617 13.1334 12.8435 13.2437 12.8243L13.2536 15.0654L12.8899 15.6773L12.9011 12.8759ZM12.8943 12.3339C12.9338 12.3045 12.9628 12.2638 12.9805 12.2168C13.0529 12.2088 13.1243 12.2002 13.1939 12.1895C13.2087 12.2038 13.2232 12.2176 13.2417 12.2274V12.2939C13.1277 12.3157 13.0135 12.336 12.9011 12.3508C12.9008 12.344 12.8953 12.3399 12.8943 12.3339ZM13.2588 16.0524L13.2642 17.1853C13.2642 17.1882 13.2642 17.191 13.2642 17.1936L12.8811 17.9953L12.8865 16.6791L13.2588 16.0524ZM12.8826 18.7505L13.1775 18.1344L13.5449 17.7328L12.8182 19.3696L12.8826 18.7505ZM13.5794 14.5182L13.5716 12.7589C13.7268 12.7238 13.8727 12.6864 14.0038 12.6477L14.0049 13.8043L13.5794 14.5182ZM14.0051 14.798L14.0059 15.643L13.5887 16.5148L13.5838 15.506L14.0051 14.798ZM13.5693 12.2228V12.2217C13.5693 12.2064 13.5565 12.196 13.5527 12.1828C13.5732 12.1615 13.5903 12.1368 13.6033 12.1093C13.663 12.095 13.7222 12.0808 13.7772 12.0647C13.8579 12.0423 13.9309 12.0174 14.0038 11.9938V12.1023C13.8566 12.1475 13.7118 12.1882 13.5693 12.2228ZM13.9981 16.4141L13.9428 16.8492L13.7024 17.3756L13.3725 17.7362L13.9981 16.4141ZM13.7609 17.6996L13.6461 18.6116L13.4784 19.0246L12.8699 19.7069L13.7609 17.6996ZM13.4309 12.0691C13.4213 12.067 13.4148 12.0584 13.4055 12.0584H13.4049C13.3795 12.0584 13.3616 12.0756 13.3395 12.0854C13.3213 12.0792 13.3037 12.0714 13.2918 12.0551C13.277 12.0338 13.2718 12.0089 13.2762 11.9821C13.284 11.9385 13.3206 11.9071 13.3798 11.9071C13.3813 11.9071 13.3829 11.9071 13.3844 11.9071C13.4345 11.9151 13.4691 11.9652 13.4602 12.0166C13.4566 12.0372 13.4452 12.0553 13.4309 12.0691ZM12.7731 12.2064C12.7619 12.2033 12.7533 12.1927 12.7406 12.1927H12.7396C12.7271 12.1927 12.7188 12.2033 12.7071 12.2064C12.6682 12.1898 12.6404 12.1516 12.6479 12.1075C12.6555 12.0634 12.6936 12.0309 12.7528 12.0309C12.7528 12.0309 12.7552 12.0309 12.7562 12.0309C12.8068 12.0395 12.8408 12.0893 12.8317 12.1413C12.8263 12.1727 12.8019 12.1947 12.7731 12.2064ZM14.0687 18.4379C14.1352 18.6531 14.3187 18.8219 14.5542 18.8517C14.6113 18.8585 14.6689 18.8572 14.7245 18.8481L13.0641 20.9112L14.0687 18.4379ZM14.577 18.6692C14.3564 18.6414 14.2001 18.4371 14.2271 18.2139C14.2406 18.1048 14.2954 18.008 14.3805 17.9413C14.4514 17.8857 14.5373 17.8564 14.6253 17.8564C14.6414 17.8564 14.6588 17.8577 14.6754 17.8595C14.8961 17.8873 15.0529 18.0908 15.0254 18.3143C15.0116 18.4239 14.9571 18.5218 14.8712 18.588C14.7876 18.6539 14.6827 18.683 14.577 18.6692ZM13.8369 18.1196L14.0295 17.1817L14.4691 16.6165L13.8369 18.1196ZM14.1471 16.1062L14.2699 15.8502L14.6848 15.2991L14.0874 16.5828L14.1471 16.1062ZM14.0916 11.7461C14.096 11.7384 14.0999 11.7313 14.1015 11.723C14.1157 11.6903 14.1448 11.6649 14.1936 11.6649H14.1972C14.2484 11.6745 14.2824 11.7238 14.2733 11.7747C14.2679 11.8048 14.2468 11.8248 14.2232 11.8383C14.204 11.831 14.1887 11.8149 14.1669 11.8149C14.1523 11.8149 14.1433 11.8266 14.1303 11.8303C14.1227 11.8243 14.1116 11.8232 14.1059 11.8142C14.0918 11.7949 14.088 11.7711 14.0916 11.7461ZM14.7684 13.5143L14.7749 14.0345L14.3335 14.9592L14.3328 14.2482L14.7684 13.5143ZM14.3317 13.2542L14.3312 12.5413C14.4231 12.5081 14.5051 12.4764 14.5737 12.4489L14.5789 12.8383L14.3317 13.2542ZM14.746 11.8113C14.7274 11.8258 14.703 11.8419 14.6954 11.8479C14.5724 11.9011 14.4511 11.9466 14.3307 11.9909V11.9787C14.3307 11.9764 14.3283 11.9754 14.3283 11.9728C14.3792 11.9375 14.4176 11.8845 14.4337 11.8209C14.5285 11.7757 14.6181 11.7272 14.7035 11.6768C14.7183 11.6921 14.7328 11.708 14.7512 11.7189C14.7505 11.7243 14.7453 11.728 14.7453 11.7339L14.746 11.8113ZM14.124 16.8834L14.7858 15.3933L14.7006 16.015L14.5661 16.3471L14.124 16.8834ZM14.4981 15.3183L14.9846 14.2877L14.8307 15.0312L14.4981 15.3183ZM14.9371 11.5722C14.927 11.5707 14.9176 11.5665 14.907 11.5678C14.8906 11.5678 14.88 11.5813 14.8647 11.586C14.818 11.5735 14.7865 11.5281 14.7951 11.4782C14.8021 11.4338 14.84 11.4016 14.899 11.4016C14.9 11.4016 14.901 11.4016 14.9023 11.4016C14.9265 11.4058 14.9478 11.4193 14.9623 11.4406C14.9768 11.4611 14.982 11.4871 14.9776 11.512C14.9737 11.5382 14.9574 11.5579 14.9371 11.5722ZM14.6303 11.4494C14.6253 11.4754 14.6287 11.5006 14.6318 11.5252C14.563 11.5662 14.488 11.6047 14.4114 11.6423C14.3759 11.5712 14.3123 11.5146 14.2274 11.4993C14.1386 11.4847 14.0597 11.5174 14.0023 11.5735C13.961 11.5418 13.907 11.5045 13.8406 11.472C14.0469 11.3827 14.1975 11.1859 14.2224 10.9518L14.48 10.8731C14.55 11.0473 14.6923 11.1774 14.8647 11.236C14.7492 11.2477 14.6492 11.3318 14.6303 11.4494ZM13.9257 11.7487C13.9241 11.7791 13.9273 11.8084 13.9358 11.8373C13.8696 11.8598 13.805 11.8822 13.731 11.9014C13.6944 11.912 13.6555 11.9209 13.6178 11.9307C13.5913 11.8357 13.515 11.7583 13.413 11.7412C13.2715 11.7173 13.1337 11.8142 13.1098 11.9541C13.1051 11.98 13.1082 12.0047 13.1113 12.0304C13.0698 12.0361 13.0285 12.0413 12.9857 12.0468C12.9706 12 12.9436 11.9606 12.9075 11.9292C13.0026 11.8926 13.0955 11.8399 13.1791 11.7552C13.2915 11.642 13.4125 11.5839 13.5384 11.5839C13.5384 11.5839 13.5397 11.5839 13.541 11.5839C13.7263 11.5854 13.8746 11.7025 13.9257 11.7487ZM12.574 12.9011L12.5612 16.2316L12.1399 16.9392L12.1417 12.9107C12.2892 12.9128 12.434 12.9097 12.574 12.9011ZM12.1417 12.4003V12.3435C12.1775 12.3196 12.2079 12.2897 12.2287 12.2511C12.2648 12.2518 12.3006 12.2524 12.3351 12.2524C12.3964 12.2524 12.4558 12.2508 12.5145 12.2487C12.5322 12.2817 12.5555 12.31 12.5856 12.3323C12.5843 12.3409 12.5763 12.3466 12.5763 12.3552V12.3822C12.4294 12.3936 12.2848 12.4001 12.1417 12.4003ZM12.5568 17.2341L12.5511 18.6547C12.5511 18.6638 12.5516 18.6729 12.5537 18.6819L12.1389 19.5477L12.1399 17.9353L12.5568 17.2341ZM12.6464 19.2453L12.576 19.9143L12.4686 20.156L11.9127 20.779L12.6464 19.2453ZM12.0558 12.201C12.0319 12.1867 12.0085 12.1688 11.9782 12.1688C11.961 12.1688 11.9493 12.1823 11.9333 12.1867C11.9148 12.1657 11.9031 12.1389 11.9086 12.1091C11.9156 12.0649 11.9527 12.032 12.0114 12.032C12.0124 12.032 12.0137 12.032 12.015 12.032C12.0662 12.0421 12.1002 12.0909 12.0911 12.1431C12.0877 12.1672 12.074 12.1867 12.0558 12.201ZM12.2245 20.7061L11.9857 21.2445C11.9151 21.2354 11.8406 21.2445 11.764 21.2697C11.7427 21.2764 11.7214 21.285 11.7006 21.2941L12.2245 20.7061ZM11.8138 12.8996L11.8118 17.4909L11.4203 18.1498L11.4234 12.8611C11.5558 12.8798 11.6856 12.8905 11.8138 12.8996ZM11.4234 12.3572V12.3087C11.4675 12.283 11.5031 12.248 11.5262 12.2023C11.6064 12.2116 11.6859 12.2197 11.764 12.2269C11.7772 12.2594 11.7964 12.2874 11.8221 12.3121C11.8208 12.3196 11.8138 12.3248 11.8138 12.3323V12.3907C11.682 12.3832 11.5522 12.3723 11.4234 12.3572ZM11.8115 18.4862L11.8107 20.1827C11.8107 20.197 11.8128 20.2097 11.8159 20.2227L11.4185 21.0542L11.4195 19.1456L11.8115 18.4862ZM11.4224 21.8239C11.5171 21.636 11.6666 21.4932 11.8208 21.4428C11.8798 21.4231 11.9667 21.4093 12.048 21.4511C12.2375 21.55 12.2715 21.8854 12.1215 22.1842C12.0267 22.3722 11.878 22.515 11.7232 22.5653C11.6638 22.5843 11.5773 22.5981 11.4953 22.557C11.3066 22.4599 11.2731 22.1238 11.4224 21.8239ZM12.2193 23.981C12.2816 24.0692 12.3458 24.1339 12.3946 24.1764C12.3008 24.2891 12.2347 24.4306 12.2079 24.5804L12.2193 23.981ZM9.99117 25.7317L9.98805 25.2138L10.5099 24.5643L9.99117 25.7317ZM10.6628 24.6707L10.5109 25.6032L10.1215 25.8893L10.6628 24.6707ZM10.4691 25.8626L10.3481 26.6035L9.99688 26.996L9.99324 26.2112L10.4691 25.8626ZM10.486 27.2597C10.5732 27.4552 10.7461 27.6066 10.9712 27.6354C10.9979 27.6385 11.0244 27.6401 11.0509 27.6401C11.1988 27.6401 11.3369 27.5876 11.4444 27.4986C11.4452 27.504 11.4483 27.5084 11.4486 27.5139C11.4465 27.5162 11.4431 27.5154 11.4405 27.5178L9.35851 30.0367L10.486 27.2597ZM10.9935 27.4576C10.7536 27.4272 10.5838 27.2073 10.6056 26.9653L10.6981 26.7395C10.7798 26.6299 10.9042 26.5554 11.0472 26.5554C11.0659 26.5554 11.0849 26.557 11.1036 26.5593C11.3491 26.5889 11.5233 26.8156 11.4924 27.0643C11.4626 27.3083 11.244 27.4916 10.9935 27.4576ZM11.1259 26.3805C11.0184 26.3667 10.9185 26.3895 10.8245 26.4269L11.6716 24.34L11.6446 26.8439C11.5797 26.6035 11.385 26.4116 11.1259 26.3805ZM10.6254 26.0474L10.9569 24.0105L11.9634 22.7541L10.6254 26.0474ZM11.2521 23.3467L11.5137 22.758C11.5818 22.7741 11.6534 22.7725 11.7263 22.7556L11.2521 23.3467ZM11.8606 23.8753L12.0462 23.4174L11.9828 26.8073C11.9288 26.8148 11.8774 26.8262 11.8286 26.8467L11.8606 23.8753ZM11.3398 12.1628C11.3149 12.1475 11.291 12.1301 11.2596 12.1301C11.2463 12.1301 11.2383 12.1413 11.2258 12.1433C11.2243 12.1413 11.2209 12.1415 11.2191 12.1387C11.2043 12.1179 11.1988 12.0927 11.2032 12.0662C11.2105 12.0229 11.2471 11.9909 11.3063 11.9909C11.3081 11.9909 11.3086 11.9909 11.3107 11.9909C11.3346 11.9954 11.3554 12.0089 11.3702 12.0288C11.385 12.0501 11.3902 12.0761 11.386 12.1018C11.3819 12.1293 11.3626 12.149 11.3398 12.1628ZM11.3473 22.6796L11.0734 23.2956L11.2032 22.4971C11.2367 22.57 11.285 22.6326 11.3473 22.6796ZM11.0958 12.2934V12.3087C10.9491 12.2827 10.8032 12.2521 10.6612 12.2163L10.6615 12.0623C10.6615 12.0592 10.6586 12.0574 10.6581 12.054C10.6667 12.0486 10.6768 12.0444 10.6853 12.0379C10.8081 12.0714 10.9252 12.0995 11.0397 12.1231C11.0457 12.1625 11.0579 12.201 11.081 12.2345C11.0885 12.2451 11.0994 12.2516 11.1082 12.2615C11.1064 12.2724 11.0958 12.2812 11.0958 12.2934ZM10.6594 12.7238C10.7983 12.7537 10.9449 12.7807 11.0955 12.8064L11.0924 18.7004L10.6412 19.4595L10.6594 12.7238ZM11.0921 19.698L11.0908 21.7411L10.6324 22.7001L10.6386 20.4602L11.0921 19.698ZM10.3035 23.387L9.91952 24.191L9.92523 21.6598L10.3094 21.0127L10.3027 23.3743C10.3027 23.3787 10.3027 23.3831 10.3035 23.387ZM10.3338 12.0605V12.1231C10.2035 12.0828 10.0745 12.0387 9.94859 11.9907L9.94963 11.8196C9.96391 11.8131 9.97689 11.8048 9.99013 11.796C10.087 11.8367 10.1853 11.8749 10.2853 11.9107C10.2985 11.951 10.3227 11.9863 10.3531 12.0161C10.3476 12.0322 10.3338 12.0434 10.3338 12.0605ZM10.3323 12.6488L10.3118 20.0132L9.92704 20.6604L9.94703 12.5403C10.0698 12.5782 10.1963 12.6145 10.3323 12.6488ZM9.86578 11.668C9.84086 11.6522 9.81697 11.634 9.78504 11.634H9.78452C9.77933 11.634 9.77751 11.6392 9.77232 11.6392C9.75622 11.6184 9.74688 11.5925 9.75207 11.5642C9.75934 11.5195 9.79672 11.4865 9.85539 11.4865C9.85669 11.4865 9.85799 11.4865 9.85903 11.4865C9.90965 11.4961 9.94366 11.546 9.93457 11.5969C9.92912 11.634 9.9003 11.6584 9.86578 11.668ZM9.59605 22.2125L9.58956 24.765C9.58956 24.7922 9.59683 24.8182 9.60773 24.841L9.20249 25.6891L9.20042 22.8789L9.59605 22.2125ZM9.19964 21.8834L9.19159 12.2622C9.32165 12.3206 9.46443 12.3772 9.61967 12.4325L9.59839 21.2123L9.19964 21.8834ZM9.36474 26.6725L9.67341 26.4459L9.28297 27.3244L9.36474 26.6725ZM9.81152 26.5874L9.81515 27.1995L9.26817 27.8104L9.81152 26.5874ZM9.80503 25.442L9.80867 26.1188L9.39616 26.4215L9.46028 25.9095L9.50571 25.8153L9.80503 25.442ZM9.62149 11.7978V11.8573C9.47534 11.7934 9.33074 11.7259 9.19159 11.6532V11.4253C9.19678 11.4219 9.20327 11.4198 9.20872 11.4157C9.33982 11.4886 9.46599 11.5551 9.58956 11.6153C9.59605 11.6641 9.61396 11.708 9.64408 11.7454C9.63603 11.7633 9.62149 11.7776 9.62149 11.7978ZM8.10385 27.9877L7.709 28.8137L7.71627 25.3745L8.10671 24.7174L8.09503 27.9331C8.09529 27.9518 8.0984 27.9692 8.10385 27.9877ZM8.15629 10.9489V11.0099C8.01455 10.9076 7.87774 10.8027 7.74742 10.6971L7.74794 10.4198C7.75053 10.4188 7.75313 10.4167 7.75495 10.4157C7.87333 10.5125 8.00027 10.6135 8.1363 10.7153C8.12981 10.7799 8.14435 10.843 8.17992 10.8928C8.17161 10.9123 8.15629 10.9274 8.15629 10.9489ZM7.98106 11.4668C8.02727 11.4917 8.08802 11.5268 8.15396 11.5732L8.10983 23.7162L7.71782 24.3748L7.7443 11.7043L7.98106 11.4668ZM7.67655 10.262C7.64929 10.2423 7.62021 10.2241 7.58465 10.2241H7.58439C7.58127 10.2241 7.57894 10.2272 7.57582 10.2277C7.5631 10.2075 7.55661 10.1836 7.56102 10.1587C7.56803 10.1151 7.60594 10.0826 7.66383 10.0826C7.66539 10.0826 7.66694 10.0826 7.6685 10.0826C7.69342 10.0868 7.71471 10.1003 7.72925 10.121C7.7443 10.1413 7.74898 10.1672 7.74456 10.1927C7.73807 10.2285 7.70978 10.2521 7.67655 10.262ZM8.62617 28.8044L8.09633 29.9962C8.03921 29.9848 7.98184 29.9734 7.92187 29.9734C7.76559 29.9734 7.61788 30.0198 7.47795 30.0873L8.62617 28.8044ZM8.16616 30.2906C8.18693 30.2997 8.20925 30.3051 8.22794 30.3166C8.56387 30.5159 8.62617 31.0398 8.36787 31.4863C8.31206 31.5821 8.22924 31.6963 8.15032 31.6963C8.13656 31.6963 8.12099 31.6888 8.10619 31.6841C7.96678 31.573 7.83127 31.4583 7.70017 31.3412L8.16616 30.2906ZM8.3619 30.0904C8.33464 30.0744 8.30245 30.066 8.27208 30.0531L8.94652 28.5349L8.7012 30.4933C8.62955 30.3262 8.5187 30.1834 8.3619 30.0904ZM9.17472 28.1907L10.237 27.0035L8.91667 30.2548L9.17472 28.1907ZM9.12825 30.6032L11.5148 27.7153C11.5449 27.7732 11.5825 27.8265 11.6285 27.8721L8.96599 31.0004L9.12825 30.6032ZM11.6228 27.3654C11.6503 27.1395 11.8403 26.9692 12.0644 26.9692C12.0836 26.9692 12.1023 26.9708 12.1217 26.9731C12.3665 27.0027 12.5405 27.2294 12.5106 27.4788C12.4813 27.7236 12.2559 27.9077 12.0122 27.8729C11.7666 27.8415 11.5924 27.6125 11.6228 27.3654ZM12.2419 26.8239C12.2167 26.8163 12.1931 26.8057 12.1666 26.801L12.2032 24.8395C12.2588 25.2289 12.5906 25.5277 12.9906 25.5277C13.3912 25.5277 13.7344 25.1977 13.7769 24.7883C13.88 24.78 14.0597 24.7616 14.2232 24.7276L12.2419 26.8239ZM13.6838 24.6115C13.6355 24.6146 13.5981 24.6549 13.5981 24.7034C13.5981 25.051 13.3198 25.3454 12.9906 25.3454C12.6534 25.3454 12.3787 25.0679 12.3787 24.7247C12.3787 24.5391 12.4608 24.3514 12.5921 24.2356C12.6139 24.2167 12.6261 24.1876 12.623 24.1588C12.6204 24.13 12.6043 24.1032 12.5794 24.0882C12.5773 24.0874 12.3847 23.9649 12.2601 23.6827C12.2526 23.6661 12.2398 23.6523 12.2253 23.6432L12.2367 22.9997C12.2367 22.989 12.2297 22.9823 12.2266 22.9729L12.2606 22.8885C12.2754 23.3875 12.5054 23.8613 12.8995 24.1694C12.6588 24.2089 12.4665 24.4051 12.4345 24.6562C12.3966 24.9649 12.6147 25.2478 12.9208 25.2868C12.9441 25.2891 12.9675 25.2907 12.9904 25.2907C13.2718 25.2907 13.5096 25.0778 13.5443 24.7964C13.5594 24.6801 13.5376 24.5643 13.4849 24.4625C13.6259 24.5012 13.771 24.5212 13.9171 24.5212C14.1355 24.5212 14.346 24.4732 14.5415 24.3919L14.4685 24.4695C14.4496 24.4664 14.4299 24.4703 14.4117 24.4802C14.2619 24.5604 13.8359 24.6024 13.6838 24.6115ZM13.3437 24.1894C13.1503 24.1165 13.0527 23.8974 13.1256 23.7011C13.1793 23.5536 13.3203 23.4555 13.4761 23.4555C13.5168 23.4555 13.5568 23.4623 13.5957 23.4753C13.617 23.4867 13.6394 23.4958 13.6617 23.5048C13.8198 23.5962 13.8899 23.7938 13.8255 23.968C13.7554 24.1575 13.5342 24.2616 13.3437 24.1894ZM13.739 23.3374C13.7198 23.3268 13.6986 23.3169 13.6716 23.307C13.4436 23.1829 13.3395 22.907 13.4283 22.6591C13.4389 22.6393 13.4488 22.618 13.4584 22.5921C13.5781 22.3662 13.8325 22.2476 14.1043 22.3486C14.1209 22.3577 14.1388 22.3652 14.157 22.373C14.3907 22.4955 14.4986 22.7738 14.4044 23.0316C14.3956 23.0482 14.3873 23.0666 14.379 23.0856C14.264 23.3125 13.9849 23.4272 13.739 23.3374ZM13.8385 21.6497C13.7782 21.8125 13.7975 21.9882 13.8779 22.1305C13.7315 22.1388 13.5939 22.192 13.4818 22.2819C13.4439 22.0978 13.3167 21.9366 13.1318 21.8665C13.0682 21.8421 13.0018 21.8307 12.9348 21.8307C12.8961 21.8307 12.8564 21.8346 12.8188 21.8429C13.1326 21.4921 13.6129 21.3006 14.0898 21.3561C13.9768 21.4202 13.8857 21.5228 13.8385 21.6497ZM12.7731 21.6271L12.8294 21.4895C12.832 21.4872 12.8359 21.488 12.8382 21.4849L14.9979 18.8024C15.0137 18.8639 15.041 18.9218 15.0755 18.9743L13.2001 21.3304C13.0436 21.4065 12.8995 21.5069 12.7731 21.6271ZM14.6988 17.677C14.5744 17.6614 14.4535 17.6962 14.345 17.7575L15.345 15.2957L15.2157 18.1651C15.2157 18.1672 15.2157 18.1685 15.2157 18.1705C15.2126 18.1729 15.2097 18.1747 15.2068 18.177C15.1684 17.9208 14.9662 17.711 14.6988 17.677ZM14.8491 15.6773L14.9101 15.1293L15.3496 14.4149L14.8491 15.6773ZM14.9732 14.7251C14.9732 14.7251 15.1458 13.9751 15.1456 13.9676L15.7774 12.6981C15.779 12.6978 15.7811 12.6978 15.7826 12.6978L15.8299 12.6799L14.9732 14.7251ZM15.0275 11.2731C15.0199 11.2682 15.0103 11.2672 15.0025 11.263C15.0137 11.2641 15.0259 11.2682 15.0378 11.2685C15.0345 11.27 15.0306 11.2716 15.0275 11.2731ZM12.1586 10.8412C12.15 10.8412 12.1402 10.8406 12.1311 10.8391C12.0428 10.823 11.9836 10.7371 11.9979 10.6491C12.0127 10.5694 12.0794 10.5128 12.1575 10.5128C12.1666 10.5128 12.1762 10.5138 12.1856 10.5156C12.2739 10.5312 12.3328 10.6169 12.3177 10.7054C12.3032 10.7843 12.2367 10.8412 12.1586 10.8412ZM10.2066 10.4375C9.93898 10.4375 9.70768 10.6093 9.62149 10.8547L9.24974 10.6231C9.32373 10.5185 9.36397 10.3918 9.36397 10.2612C9.36397 9.91594 9.08645 9.63583 8.74455 9.63583C8.51818 9.63583 8.31309 9.7594 8.20484 9.95385C8.13345 9.87596 7.98392 9.69995 7.67681 9.30276C7.8414 9.13558 7.9198 8.95567 7.90863 8.76746C7.89435 8.52032 7.72873 8.32665 7.62099 8.22619L7.98911 7.92998C8.06439 8.19971 8.30998 8.39805 8.60047 8.39805C8.73573 8.39805 8.86501 8.35547 8.97196 8.27785L9.14746 8.50422C9.05711 8.61689 9.00857 8.75656 9.00857 8.90505C9.00857 9.25863 9.29309 9.54601 9.64278 9.54601C9.76453 9.54601 9.88473 9.50992 9.98623 9.44347L10.2692 9.64674C10.237 9.72488 10.2204 9.80899 10.2204 9.89414C10.2204 10.2475 10.5052 10.5348 10.8554 10.5348C11.0449 10.5348 11.223 10.4499 11.3429 10.3048L11.5968 10.3965C11.5807 10.4531 11.5729 10.512 11.5729 10.5701C11.5729 10.7166 11.622 10.8521 11.7048 10.9606C11.4366 10.9915 11.2201 11.1966 11.1682 11.4606L10.7643 11.3349C10.8045 11.25 10.8258 11.1574 10.8258 11.0626C10.8258 10.7176 10.5483 10.4375 10.2066 10.4375ZM10.3637 10.9837C10.3484 11.072 10.2705 11.133 10.1765 11.1171C10.0893 11.1013 10.0304 11.0159 10.0462 10.9276C10.06 10.8484 10.1269 10.7924 10.2043 10.7924C10.2139 10.7924 10.2233 10.7926 10.2321 10.7944C10.3201 10.8092 10.3795 10.8954 10.3637 10.9837ZM8.93225 10.1955C8.91953 10.2721 8.85151 10.3305 8.77415 10.3305C8.76506 10.3305 8.75598 10.3297 8.74663 10.3282C8.65862 10.3121 8.60047 10.2269 8.61605 10.1369C8.62877 10.0587 8.69523 10.0019 8.77415 10.0019C8.78375 10.0019 8.79284 10.0026 8.80244 10.0039C8.89123 10.0203 8.94912 10.1039 8.93225 10.1955ZM11.9701 11.498C11.9553 11.5854 11.8722 11.6467 11.784 11.6306C11.6957 11.615 11.6376 11.5296 11.6534 11.4411C11.6669 11.3614 11.7339 11.3051 11.8118 11.3051C11.8208 11.3051 11.8299 11.3059 11.8395 11.3074C11.927 11.3227 11.9857 11.4084 11.9701 11.498ZM6.96965 9.47955C6.96654 9.49565 6.95641 9.50681 6.94706 9.51875C6.93668 9.5229 6.92552 9.5229 6.91669 9.53095C6.91228 9.53511 6.91332 9.54056 6.91046 9.54627C6.89982 9.55042 6.89021 9.55795 6.87853 9.55795C6.86503 9.5525 6.85439 9.54263 6.83933 9.54186C6.80402 9.52239 6.77936 9.48812 6.78689 9.44606C6.79364 9.40193 6.83102 9.36948 6.89021 9.36948C6.89021 9.36948 6.89255 9.36948 6.89436 9.36948C6.91851 9.37337 6.93954 9.38687 6.95381 9.40738C6.96835 9.42763 6.97432 9.45359 6.96965 9.47955ZM5.26406 33.1166C5.28431 32.9995 5.35492 32.8931 5.45357 32.8329C5.49303 32.8077 5.5491 32.7841 5.61608 32.7841C5.64178 32.7841 5.66852 32.788 5.69682 32.7963C5.74718 32.8108 5.77418 32.8215 5.79028 32.8267L5.71577 32.9827C5.70149 33.0079 5.69033 33.0336 5.67813 33.0619L5.45772 33.5219C5.31936 33.443 5.23421 33.2849 5.26406 33.1166ZM5.34687 33.7555L5.2934 33.8666C5.2934 33.8225 5.29132 33.7768 5.28742 33.729C5.30508 33.7412 5.3274 33.7456 5.34687 33.7555ZM5.46421 28.1647C3.7986 25.0396 3.73682 21.8042 3.7363 21.7647C3.55068 17.5659 5.19786 14.601 6.66202 12.8217L6.65605 26.1624L5.46421 28.1647ZM6.65527 27.1585L6.65423 30.26C6.27469 29.8067 5.94033 29.3345 5.64801 28.8521L6.65527 27.1585ZM4.12233 26.3387C4.12389 26.3441 4.12544 26.3478 4.127 26.3524L3.68723 27.3646L2.69789 26.9365L3.3186 26.151L4.12233 26.3387ZM3.46242 25.9674L3.84974 25.4783C3.90816 25.6829 3.97046 25.8906 4.04029 26.1035L3.46242 25.9674ZM3.79679 27.6416L4.64413 28.009C4.30976 28.0697 3.81807 28.1907 3.46216 28.4129L3.79679 27.6416ZM3.88064 27.4482L4.23084 26.6421C4.37336 27.0269 4.53458 27.4178 4.71967 27.8119L3.88064 27.4482ZM3.22774 21.5921L2.39441 21.2232L3.2267 20.719C3.21917 21.0041 3.21865 21.2946 3.22774 21.5921ZM2.71398 19.8367L3.31418 19.4883C3.28589 19.7318 3.26538 19.9813 3.24928 20.2349L2.71398 19.8367ZM3.23734 22.002C3.24747 22.1754 3.26772 22.4454 3.30873 22.7907L2.49773 22.4675L3.23734 22.002ZM3.35858 23.1702C3.39051 23.3937 3.42997 23.6372 3.47877 23.8997L2.62909 23.6616L3.35858 23.1702ZM2.91829 18.5605L3.51226 18.2453C3.46086 18.4888 3.42062 18.7443 3.38142 19L2.91829 18.5605ZM3.56029 24.3096C3.60909 24.5401 3.66517 24.7811 3.72903 25.0305L2.95074 24.8106L3.56029 24.3096ZM3.96579 16.6121L3.06678 17.074L2.57977 16.4564L3.53407 16.0213L3.96579 16.6121ZM3.72981 15.932L4.3487 15.6493C4.24616 15.8845 4.14751 16.1251 4.05457 16.3769L3.72981 15.932ZM3.87519 16.8961C3.78303 17.1812 3.70177 17.4782 3.62571 17.784L3.1997 17.2432L3.87519 16.8961ZM4.56209 15.1737L4.21604 14.5976L5.026 14.3009C4.86635 14.5758 4.71137 14.8645 4.56209 15.1737ZM5.21837 13.9805L4.96837 13.256L5.93539 12.9247C5.69786 13.2396 5.4572 13.5917 5.21837 13.9805ZM4.89776 13.0517L4.70721 12.4987C4.88919 12.6335 5.14594 12.7768 5.46551 12.8432L4.89776 13.0517ZM4.76848 13.3235L5.0273 14.0753L4.10416 14.413L3.68827 13.7217L3.68983 13.7202L4.76848 13.3235ZM3.90322 14.487L3.12987 14.77C3.28199 14.4429 3.42997 14.1623 3.56652 13.9258L3.90322 14.487ZM4.01407 14.6711L4.43852 15.3772L3.60208 15.7588L3.06237 15.0195L4.01407 14.6711ZM3.4066 15.8471L2.55718 16.2344C2.68024 15.8588 2.80822 15.513 2.93724 15.2048L3.4066 15.8471ZM2.87572 17.1724L2.19893 17.5207C2.27396 17.2017 2.35495 16.9016 2.43932 16.6178L2.87572 17.1724ZM3.00889 17.3406L3.52758 17.9984L2.75708 18.4068L2.1304 17.8112C2.13273 17.7996 2.13559 17.7863 2.13741 17.7746C2.1522 17.7741 2.16752 17.7723 2.18102 17.7655L3.00889 17.3406ZM2.56056 18.5116L1.94115 18.8403C1.98398 18.569 2.03071 18.3081 2.08185 18.0573L2.56056 18.5116ZM2.72177 18.6638L3.32613 19.238L2.53018 19.6993L1.88429 19.2185C1.89026 19.1749 1.89727 19.1339 1.9035 19.0903C1.90922 19.0882 1.91571 19.0913 1.92168 19.0882L2.72177 18.6638ZM2.33185 19.8151L1.77915 20.1355C1.80044 19.9042 1.82381 19.6767 1.85106 19.4576L2.33185 19.8151ZM2.51617 19.9519L3.21969 20.4763L2.15947 21.1191L1.71036 20.9197C1.72412 20.6749 1.73969 20.496 1.75112 20.3961L2.51617 19.9519ZM1.93985 21.2512L1.69297 21.4018C1.69478 21.3104 1.69764 21.2268 1.70049 21.1456L1.93985 21.2512ZM2.17505 21.3556L3.16777 21.7974L2.25449 22.3717L1.69452 22.1489C1.68959 21.9724 1.68855 21.8068 1.68907 21.6508L2.17505 21.3556ZM2.03356 22.51L1.71971 22.7071C1.7127 22.5944 1.70698 22.4851 1.70257 22.3784L2.03356 22.51ZM2.27577 22.6066L3.24539 22.9921L2.36222 23.5869L1.77864 23.4241C1.76176 23.2606 1.7467 23.1009 1.73502 22.9464L2.27577 22.6066ZM2.13273 23.7421L1.84146 23.9376C1.82822 23.8403 1.81654 23.7452 1.80563 23.651L2.13273 23.7421ZM2.39909 23.8158L3.46501 24.1147L2.70282 24.7413L1.9331 24.5238C1.91181 24.4028 1.89442 24.2857 1.87651 24.1686L2.39909 23.8158ZM2.50474 24.9044L2.08341 25.252C2.04551 25.0845 2.01046 24.9205 1.97905 24.7561L2.50474 24.9044ZM2.75189 24.9742L3.75577 25.2579L3.2363 25.915L2.18284 25.6679C2.16648 25.6048 2.15272 25.5432 2.13741 25.4799L2.75189 24.9742ZM3.09222 26.0967L2.5982 26.7218C2.54005 26.5295 2.44945 26.3278 2.31108 26.1188C2.28953 26.0458 2.26954 25.9721 2.24981 25.8989L3.09222 26.0967ZM2.67764 27.1569L3.60338 27.5578C3.60338 27.5578 3.05276 28.8407 3.05588 28.8599L2.80095 29.4669C2.38429 28.7369 2.33444 28.4583 2.32847 28.409C2.41258 28.2911 2.49384 28.1548 2.55614 27.9923C2.62831 27.802 2.70178 27.5084 2.67764 27.1569ZM3.25603 28.917C3.44658 28.4669 4.40971 28.2486 4.9014 28.1816C5.30845 28.9825 5.81701 29.7787 6.44681 30.5219C6.35984 30.5385 6.27599 30.5712 6.20097 30.6232C6.05793 30.7236 5.96291 30.8734 5.93228 31.0468C5.90294 31.2148 5.94033 31.3799 6.02548 31.5123C5.90502 31.5177 5.7882 31.5572 5.68669 31.6286C5.54261 31.7283 5.44682 31.8788 5.41593 32.053C5.38607 32.2272 5.42735 32.3952 5.52003 32.5307C5.44942 32.5429 5.3814 32.5702 5.3165 32.6099C5.19423 32.6852 5.09843 32.8023 5.04392 32.937C4.86583 32.5613 4.57456 32.0845 4.10519 31.438C3.57379 30.7073 3.20204 30.1413 2.94062 29.7062L3.25603 28.917ZM5.67397 32.0982C5.6924 31.9941 5.74978 31.9035 5.83648 31.8435C5.9019 31.7979 5.97641 31.775 6.05351 31.775C6.07584 31.775 6.09842 31.7766 6.12127 31.7805C6.24146 31.8017 6.26846 31.8215 6.27054 31.823L5.9328 32.5289C5.75263 32.4565 5.64178 32.2846 5.67397 32.0982ZM6.1898 31.0928C6.20901 30.9887 6.26587 30.8981 6.35102 30.8386C6.41644 30.7922 6.49198 30.7693 6.5696 30.7693C6.59193 30.7693 6.61477 30.7709 6.6371 30.7748C6.65319 30.7779 6.66358 30.7815 6.67604 30.7862C6.69421 30.8059 6.71368 30.8256 6.73263 30.8461C6.73263 30.8508 6.72978 30.8576 6.72874 30.8635L6.60802 31.1161C6.51976 31.26 6.44681 31.4129 6.40398 31.5138C6.25107 31.4396 6.15917 31.2685 6.1898 31.0928ZM7.29909 6.09225C7.48678 5.97024 7.63735 5.93649 7.7077 5.93649C7.72847 5.93649 7.74249 5.93909 7.74846 5.94402C7.75884 5.97855 7.67733 6.1434 7.43226 6.30409C7.19083 6.46219 7.00885 6.47335 6.98341 6.45258C6.97328 6.41754 7.0535 6.25295 7.29909 6.09225ZM5.67631 7.43206C5.86296 7.30953 6.01353 7.27526 6.08466 7.27526C6.10517 7.27526 6.11945 7.27812 6.12542 7.28253C6.13607 7.31784 6.05403 7.48243 5.81 7.64182C5.56624 7.80278 5.384 7.80953 5.35985 7.7911C5.34921 7.75605 5.43073 7.59198 5.67631 7.43206ZM5.01199 8.15895L5.25783 7.97385C5.27834 7.98502 5.30196 7.99462 5.33156 8.00085C5.48239 8.03019 5.7085 7.96165 5.92501 7.81861C6.19629 7.64182 6.41124 7.37002 6.3113 7.18518L7.00158 6.66805C7.14748 6.67974 7.35179 6.6086 7.54753 6.4801C7.76923 6.3342 7.95406 6.12496 7.95848 5.95207L8.0444 5.88769L8.33256 6.0759L5.48031 8.37676L5.01199 8.15895ZM5.63503 8.95178C5.59505 8.98163 5.52029 9.04056 5.41203 9.13064C5.31962 8.95879 5.17164 8.66829 5.03743 8.40246L5.39438 8.56835C5.39879 8.58756 5.39438 8.60677 5.40606 8.62312C5.41048 8.63195 5.49251 8.75812 5.63503 8.95178ZM5.83389 9.21242C6.04027 9.47332 6.31805 9.79756 6.6628 10.147V10.7321C6.50808 10.4128 6.21239 10.1685 5.84038 10.1029C5.21136 9.98967 4.62881 10.4198 4.52108 11.0353C4.41204 11.6605 4.82585 12.2573 5.44448 12.3669C5.51094 12.3788 5.5774 12.3845 5.6436 12.3845C6.08804 12.3845 6.47589 12.1239 6.66306 11.7368V12.0361C6.49847 12.2186 6.32272 12.4292 6.14126 12.6581C6.12646 12.6537 6.11374 12.6462 6.09661 12.6485C6.07117 12.6527 6.0465 12.6522 6.02106 12.655C6.00886 12.655 5.99666 12.6493 5.98446 12.6529L5.96758 12.6597C5.92034 12.6636 5.87309 12.6711 5.82584 12.6711C5.04963 12.6711 4.55924 12.0828 4.55509 12.0764C4.54963 12.0696 4.53977 12.068 4.53302 12.0631C4.52419 12.0558 4.51692 12.0504 4.50628 12.047C4.49096 12.0418 4.47694 12.0421 4.46111 12.0449C4.45358 12.046 4.44683 12.0405 4.4393 12.0436C4.43619 12.0449 4.43567 12.0475 4.43385 12.0491C4.43074 12.0499 4.42788 12.0491 4.42477 12.0499C4.41049 12.0577 4.40296 12.0719 4.39335 12.0841C4.3892 12.0901 4.38219 12.0932 4.37933 12.0995C4.36869 12.1228 4.36532 12.1485 4.37362 12.1732L4.37492 12.1766L4.37518 12.1786L4.66775 13.1355L3.88064 13.4237C4.07768 13.1319 4.20852 12.9829 4.22565 12.9642C4.26978 12.9497 4.30639 12.9185 4.31547 12.8723C4.32404 12.8271 4.51381 11.8168 3.79367 11.3087C4.79781 10.0221 5.68617 9.32353 5.83389 9.21242ZM6.55636 11.3975C6.4681 11.9009 5.99925 12.2526 5.48057 12.1597C4.9772 12.0701 4.63946 11.5823 4.7285 11.072C4.80664 10.6221 5.19007 10.2957 5.64126 10.2957C5.69448 10.2957 5.74848 10.3009 5.80351 10.3105C6.30792 10.3996 6.64567 10.8876 6.55636 11.3975ZM3.15064 11.6335C3.17478 11.5237 3.25837 11.4123 3.39155 11.4123C3.44866 11.4123 3.51018 11.4315 3.57664 11.4702C3.93074 11.676 4.04471 12.0408 4.07404 12.3391C3.88946 12.1153 3.65868 12.0286 3.44554 11.9484C3.37623 11.9224 3.30692 11.8962 3.2389 11.8658C3.14596 11.8238 3.13246 11.7186 3.15064 11.6335ZM3.19451 12.1306C3.24747 12.1522 3.30017 12.1747 3.35287 12.1947C3.63402 12.2999 3.87752 12.3962 4.02653 12.7929C3.97176 12.8531 3.681 13.1843 3.30718 13.8344C3.28693 13.7342 3.21086 13.6545 3.11066 13.636C3.26408 13.5883 3.39336 13.4878 3.47955 13.3494C3.67633 13.033 3.58209 12.6132 3.26772 12.4138C3.19866 12.3697 3.12779 12.3474 3.05692 12.3375C3.10313 12.2669 3.14856 12.1979 3.19451 12.1306ZM2.79472 12.7552C2.83029 12.737 2.86767 12.7248 2.90998 12.7248C3.05432 12.7248 3.17088 12.843 3.17088 12.9876C3.17088 13.1335 3.05432 13.2521 2.90998 13.2521C2.77006 13.2521 2.65765 13.1381 2.65272 12.9974C2.70022 12.9154 2.74721 12.8341 2.79472 12.7552ZM2.53096 13.2095C2.60755 13.3432 2.74566 13.4359 2.90998 13.4359C3.15479 13.4359 3.35442 13.2355 3.35442 12.9883C3.35442 12.7469 3.16258 12.5527 2.92556 12.5444C2.93101 12.5361 2.93594 12.5275 2.9414 12.5192C3.01538 12.5021 3.09145 12.5192 3.17114 12.5699C3.39933 12.715 3.46865 13.0211 3.32405 13.2529C3.25474 13.3655 3.14622 13.4429 3.0185 13.472C2.89285 13.5011 2.76149 13.4782 2.65194 13.4086C2.57899 13.3624 2.54057 13.3006 2.5185 13.2321C2.5224 13.2241 2.52681 13.2171 2.53096 13.2095ZM2.02785 14.1937L2.06576 14.2087L2.00397 14.243C2.01124 14.2261 2.02007 14.2103 2.02785 14.1937ZM2.24047 14.3214L2.4048 14.5888L2.06576 14.797L1.91389 14.5034L2.24047 14.3214ZM2.50059 14.7448L2.51643 14.77C2.45802 14.9161 2.38922 15.0916 2.31627 15.2809L2.15091 14.9605L2.50059 14.7448ZM2.46554 14.3549L2.67452 14.4323L2.55796 14.5096L2.46554 14.3549ZM1.80563 14.6937L1.9087 14.8933L1.66389 15.0431C1.71114 14.9252 1.75735 14.8076 1.80563 14.6937ZM1.99359 15.0566L2.20049 15.4574L1.7737 15.7526L1.56706 15.3178L1.99359 15.0566ZM2.15065 15.7157C2.09146 15.8741 2.03382 16.0343 1.98061 16.1877L1.85418 15.9211L2.15065 15.7157ZM1.47438 15.5501L1.6208 15.8585L1.29577 16.0836C1.35263 15.9016 1.4126 15.7243 1.47438 15.5501ZM1.70049 16.0265L1.89104 16.4286L1.46555 16.7069L1.25813 16.3318L1.70049 16.0265ZM1.89883 16.6165L1.82484 16.8629L1.59068 16.8009L1.65221 16.7765L1.89883 16.6165ZM1.15922 16.5317L1.26229 16.7183L1.1208 16.6778C1.13352 16.6287 1.14546 16.5792 1.15922 16.5317ZM0.694015 18.8102C0.77579 18.9104 0.813173 19.0433 0.790328 19.1757C0.770598 19.2884 0.708293 19.3852 0.620807 19.453C0.641316 19.2357 0.666497 19.0215 0.694015 18.8102ZM0.58654 19.6661C0.590693 19.6645 0.941676 19.5215 0.986068 19.1098C1.02968 18.7087 0.717899 18.5399 0.70414 18.5327L0.734773 18.3533C0.751647 18.3619 1.17714 18.6274 1.12392 19.1248C1.0707 19.6131 0.587319 19.8413 0.568108 19.8504L0.58654 19.6661ZM0.575896 20.055C0.979838 19.9644 1.29785 19.6214 1.34666 19.1949C1.39572 18.7676 1.16364 18.3611 0.790847 18.1812C0.837575 17.9068 0.889756 17.6391 0.947647 17.3785L1.65325 17.5433C1.63819 17.6051 1.61924 17.6877 1.59691 17.7879C1.58627 17.752 1.56965 17.7173 1.54785 17.6859C1.49385 17.6095 1.41311 17.5584 1.32173 17.5426C1.22932 17.5265 1.13664 17.5467 1.06031 17.5994C0.982953 17.6537 0.931292 17.7352 0.915976 17.8286C0.881968 18.0202 1.00917 18.2038 1.19946 18.238C1.28928 18.2544 1.3843 18.2349 1.46114 18.1809C1.48554 18.1633 1.50839 18.1425 1.52734 18.1202C1.41026 18.7082 1.26592 19.5895 1.23503 20.4875C1.1896 20.3821 1.09458 20.3008 0.972829 20.279C0.880929 20.2619 0.787732 20.2832 0.711149 20.3369C0.634047 20.3917 0.582386 20.4724 0.56707 20.5651C0.532542 20.7554 0.660267 20.9395 0.850036 20.9745H0.850556L0.851075 20.9753L0.547599 20.9792C0.549417 20.6651 0.559281 20.3561 0.575896 20.055ZM1.34873 17.9057C1.34354 17.9307 1.33056 17.9514 1.31031 17.9657C1.29058 17.9792 1.26748 17.9821 1.24489 17.9802C1.19713 17.9717 1.16467 17.9239 1.17402 17.8738C1.17817 17.8489 1.19115 17.8281 1.2114 17.8144C1.22646 17.8037 1.24256 17.7983 1.26021 17.7983C1.26566 17.7983 1.27111 17.799 1.2763 17.8001C1.29993 17.8037 1.31992 17.8164 1.33316 17.8362C1.34769 17.8564 1.35263 17.8808 1.34873 17.9057ZM1.11224 20.9166C1.16338 20.8816 1.20258 20.8346 1.22906 20.7798C1.22906 20.8437 1.22906 20.9067 1.23035 20.9698L0.983991 20.9729C1.02838 20.9639 1.07252 20.9449 1.11224 20.9166ZM0.961666 20.7022C0.941676 20.7159 0.920389 20.7188 0.897284 20.7167C0.8477 20.7076 0.816288 20.6612 0.825893 20.6098C0.829528 20.5859 0.842508 20.5651 0.862757 20.5513C0.880929 20.5371 0.901957 20.5337 0.927398 20.5363C0.975684 20.5451 1.00839 20.5934 1.00035 20.6419C0.995674 20.6671 0.981915 20.6876 0.961666 20.7022ZM0.554089 22.2011L0.758397 22.3852L0.565252 22.6946C0.558503 22.5038 0.555128 22.339 0.554089 22.2011ZM0.56006 21.7009L0.640537 21.6962L0.558503 21.8172C0.559801 21.7798 0.560839 21.7533 0.560839 21.7502C0.56006 21.7336 0.560579 21.7175 0.56006 21.7009ZM1.04837 24.2198L0.754762 23.9703L1.00476 23.517L1.29474 23.7961L1.04837 24.2198ZM1.34146 23.9932L1.37495 24.2473L1.20154 24.2816L1.34146 23.9932ZM0.960108 23.2195L0.644691 22.9161L0.897284 22.51L1.20154 22.7831L0.960108 23.2195ZM1.09718 23.3512L1.25294 23.0705C1.25969 23.225 1.26878 23.3787 1.27942 23.5261L1.09718 23.3512ZM0.996193 22.3519L1.20413 22.0158L1.21244 22.0225C1.21686 22.1778 1.22283 22.3631 1.22932 22.5617L0.996193 22.3519ZM1.1626 21.7891L0.996193 21.678L1.17766 21.6567L1.1626 21.7891ZM1.05798 21.9039L0.856786 22.2263L0.635864 22.0285L0.837316 21.7341L1.05798 21.9039ZM0.587578 23.1154L0.868468 23.3862L0.644431 23.7909C0.619509 23.5474 0.600818 23.3231 0.587578 23.1154ZM0.684929 24.1528L0.883266 24.3208L0.711409 24.3475C0.702582 24.2821 0.692458 24.2159 0.684929 24.1528ZM1.1639 25.6292C1.21036 25.5913 1.26592 25.5775 1.3207 25.5819C1.38015 25.5902 1.43388 25.6253 1.47075 25.6816C1.55096 25.8003 1.5302 25.969 1.42558 26.0573C1.37989 26.0959 1.32719 26.1113 1.26722 26.1053C1.20777 26.097 1.15429 26.0619 1.11743 26.0056C1.03825 25.887 1.05902 25.718 1.1639 25.6292ZM1.09692 26.2346C1.14079 26.2618 1.19038 26.2795 1.2423 26.2878C1.34977 26.3031 1.45751 26.2712 1.54447 26.198C1.71841 26.0513 1.75423 25.7735 1.62417 25.5788C1.55771 25.4791 1.4588 25.416 1.34692 25.3994C1.23996 25.3833 1.13404 25.4155 1.04837 25.4861C1.08965 25.4124 1.15637 25.36 1.26073 25.3501C1.36483 25.341 1.48658 25.3654 1.5694 25.436C1.67038 25.5204 1.73632 25.6512 1.748 25.7935C1.77552 26.132 1.58445 26.3252 1.34769 26.3472C1.27345 26.3555 1.19141 26.3275 1.14261 26.2849C1.12599 26.2709 1.11094 26.2551 1.09692 26.2346ZM1.13794 26.4924C1.18674 26.5105 1.23918 26.5227 1.29603 26.5227C1.31239 26.5227 1.32952 26.5219 1.34744 26.5204C1.6831 26.4885 1.93466 26.1598 1.90662 25.788C1.89208 25.598 1.80408 25.4215 1.66571 25.3044C1.54447 25.2024 1.3965 25.1499 1.24178 25.1692C1.07667 25.1845 0.959329 25.2712 0.886381 25.4088C0.826672 25.1045 0.775271 24.8093 0.734773 24.5295L1.46919 24.416C1.49437 24.4129 1.51618 24.3994 1.53071 24.3787C1.53331 24.374 1.53461 24.3688 1.53669 24.3641C1.48996 24.0827 1.34432 23.0796 1.34588 21.4984C1.32978 21.4862 1.31161 21.4794 1.28928 21.4802L0.556685 21.5168C0.554609 21.4241 0.553311 21.3327 0.552013 21.2408L1.35522 21.2302L1.36639 21.1837C1.36639 21.1837 1.44011 19.8034 1.53071 19.1952C1.61794 18.609 1.83445 17.48 1.87962 17.3343C1.86768 17.3268 1.85548 17.3201 1.8412 17.3167L1.00502 17.1215C1.02683 17.0306 1.05123 16.9429 1.07407 16.8538L1.81186 17.0672C1.82043 17.0698 1.829 17.0709 1.83731 17.0709C1.85392 17.0709 1.87027 17.0667 1.88507 17.0579C1.90688 17.0449 1.92297 17.0231 1.92817 16.9979C1.95309 16.8793 1.97282 16.7985 2.03097 16.6184C2.03097 16.6184 2.85287 14.4265 2.85495 14.4216C2.86429 14.399 2.86429 14.3731 2.85391 14.3502C2.84456 14.3279 2.82613 14.3095 2.80329 14.3012L2.10677 14.0273C2.20698 13.8196 2.30771 13.6197 2.41103 13.4271C2.44893 13.4764 2.49176 13.5231 2.55173 13.5623C2.66881 13.6368 2.80329 13.6727 2.9388 13.6651C2.86429 13.7087 2.81419 13.7905 2.81419 13.8832C2.81419 14.0218 2.92686 14.1345 3.06445 14.1345C3.0943 14.1345 3.12286 14.1267 3.1496 14.1176C2.7441 14.8782 2.28097 15.9813 1.93492 17.4654C1.93673 17.4548 1.93621 17.4442 1.93621 17.4335C1.86872 17.8107 1.65169 19.0384 1.6034 19.5148C1.54551 20.0815 1.46374 21.0345 1.44738 21.2974C1.44427 21.3462 1.43778 21.3795 1.42973 21.4023C1.40688 22.5126 1.491 24.313 2.07354 26.2463C2.12261 26.5962 2.12261 27.1764 1.9183 27.3644C1.82614 27.4487 1.73009 27.5173 1.64598 27.5765C1.59821 27.6107 1.5507 27.645 1.50553 27.6813C1.36379 27.2795 1.24204 26.881 1.13794 26.4924ZM1.7859 28.5115H1.78305C1.7628 28.5001 1.74255 28.5001 1.71919 28.4993C1.66571 28.4757 1.54421 28.4607 1.49048 28.2408C1.47438 28.1777 1.53487 28.0749 1.55148 28.0315C1.56342 28.0141 1.5694 27.9996 1.57459 27.9791C1.61612 27.9235 1.68829 27.8695 1.79914 27.7904C1.89 27.7265 1.99255 27.6543 2.09743 27.5578C2.27551 27.3942 2.34197 27.1029 2.3604 26.8298C2.47852 27.2444 2.417 27.604 2.32432 27.8628C2.18439 28.2522 1.94919 28.4773 1.883 28.497C1.84561 28.5071 1.81342 28.5115 1.7859 28.5115ZM4.63219 35.0836C4.08806 35.3985 3.67633 35.4608 3.4728 35.2654C3.29549 35.0935 3.30328 34.7589 3.36559 34.6418C3.46346 34.4601 3.57846 34.3557 3.70878 34.33C3.92269 34.2921 4.1392 34.4676 4.14076 34.4692C4.17425 34.4972 4.21838 34.5125 4.26199 34.5071C4.30613 34.5039 4.34585 34.4827 4.37414 34.45C5.26095 33.3799 3.54289 31.397 3.47514 31.3194C2.87649 30.5611 2.36326 29.6956 1.93803 28.7525C1.94141 28.7517 1.94426 28.7517 1.9479 28.7509C1.998 28.7379 2.05823 28.6992 2.12053 28.6468C2.24099 29.0134 2.63143 29.8537 3.89414 31.5925C5.17657 33.3586 5.08364 33.7823 4.9663 34.3176C4.94034 34.4339 4.91489 34.5564 4.89776 34.6948L4.71708 35.072C4.68826 35.0707 4.65815 35.0675 4.63219 35.0836ZM4.75991 35.5117L6.26872 32.5857C6.27262 32.5798 6.27547 32.5728 6.27807 32.566L6.92967 31.0572C6.94447 31.0717 7.0535 31.1834 7.06804 31.1979L4.99771 35.4959C4.97954 35.4972 4.77782 35.5109 4.75991 35.5117ZM5.23914 35.4577L5.60258 34.79L6.71654 32.3882L6.71576 32.389L6.72822 32.363L7.21965 31.3415C7.52442 31.6327 7.851 31.9118 8.20484 32.1727L8.2565 32.4222C8.25209 32.4746 8.24222 32.525 8.24222 32.5782C8.24222 32.6216 8.27337 32.6595 8.31725 32.668C8.35411 32.6748 8.39279 32.6818 8.43043 32.6893C7.08102 34.662 5.97978 35.2981 5.23914 35.4577ZM9.39382 32.4557C9.39278 32.4648 9.37565 32.477 9.34865 32.4884L8.66278 32.3659C8.66278 32.3651 8.66148 32.3643 8.66148 32.3635C8.68484 32.1771 8.84216 32.0356 9.02856 32.0356C9.04439 32.0356 9.05841 32.0364 9.07347 32.038C9.2757 32.0642 9.419 32.2511 9.39382 32.4557ZM9.09606 31.8565C8.85748 31.8293 8.63915 31.9622 8.53791 32.1675L8.52207 32.0946L8.67887 31.7098C8.96392 31.2535 9.46028 30.9464 10.0317 30.9432C9.86474 31.0422 9.74117 31.2057 9.71625 31.4103C9.69574 31.5821 9.74117 31.7509 9.84709 31.8879C9.86188 31.9077 9.88499 31.9183 9.90238 31.9367C9.75674 31.9762 9.6259 32.0593 9.53141 32.1802C9.45405 32.0094 9.29387 31.8817 9.09606 31.8565ZM10.8063 31.5476C10.7759 31.7971 10.5589 31.985 10.2962 31.9484C10.175 31.9339 10.067 31.8724 9.9922 31.775C9.91614 31.6761 9.88395 31.5551 9.89823 31.4321C9.92678 31.2008 10.1212 31.0261 10.3515 31.0261C10.3702 31.0261 10.3899 31.0268 10.4089 31.0292C10.6602 31.0616 10.838 31.2937 10.8063 31.5476ZM10.4317 30.4199C10.4561 30.2267 10.6184 30.0853 10.8053 30.0853C10.8206 30.0853 10.8351 30.086 10.8515 30.0876C11.0558 30.1143 11.202 30.3059 11.1773 30.5128C11.1648 30.6148 11.1142 30.7044 11.0335 30.767C10.9548 30.8285 10.8549 30.856 10.7578 30.8423C10.5524 30.8176 10.4063 30.6268 10.4317 30.4199ZM10.5854 32.6748C10.5818 32.687 10.5807 32.6976 10.5789 32.7083L9.57243 32.5281C9.6272 32.274 9.83281 32.0907 10.0745 32.0907C10.0947 32.0907 10.116 32.0922 10.1371 32.0946C10.419 32.1304 10.6202 32.3905 10.5854 32.6748ZM11.2222 32.8225L10.7593 32.7405C10.7606 32.7345 10.7635 32.7252 10.7637 32.7207C10.8061 32.5556 10.9535 32.4409 11.1212 32.4409C11.1363 32.4409 11.1508 32.4417 11.1661 32.4432C11.3494 32.4668 11.4774 32.6226 11.4831 32.8046C11.358 32.8565 11.2521 32.8316 11.2222 32.8225ZM11.6495 32.6566C11.5893 32.4505 11.413 32.2885 11.1887 32.2605C11.0195 32.2407 10.859 32.2971 10.7419 32.4087C10.7064 32.2893 10.6371 32.1852 10.5483 32.1C10.7796 32.024 10.9569 31.8241 10.9888 31.5699C11.0052 31.4383 10.9706 31.3165 10.9159 31.2065C11.3611 31.497 11.6576 32.0014 11.6576 32.5761C11.6578 32.6042 11.6521 32.6299 11.6495 32.6566ZM11.7313 31.9515C11.5877 31.5538 11.3089 31.2215 10.9538 31.01C11.0226 30.9902 11.0885 30.9575 11.1467 30.9118C11.2656 30.8192 11.3408 30.6852 11.359 30.5346C11.3961 30.2273 11.1786 29.9445 10.8743 29.9048H10.8733C10.568 29.8721 10.2876 30.0873 10.2487 30.397C10.2316 30.5377 10.2689 30.6738 10.3421 30.7833C9.9787 30.7218 9.58385 30.7802 9.26038 30.9378L9.4216 30.7478C9.43562 30.7478 9.45041 30.7446 9.46365 30.7371C9.615 30.6535 9.97221 30.6063 10.0971 30.5949C10.1451 30.5902 10.1807 30.5507 10.1807 30.5035C10.1807 30.1551 10.459 29.8599 10.7879 29.8599C11.1246 29.8599 11.399 30.1374 11.399 30.4799C11.399 30.667 11.3172 30.8539 11.1861 30.9697C11.162 30.9902 11.1513 31.0229 11.1568 31.0541C11.1615 31.0852 11.1827 31.112 11.2121 31.1242C11.2248 31.1296 11.5301 31.2602 11.718 31.7174C11.7227 31.728 11.7292 31.7379 11.7367 31.7462L11.7313 31.9515ZM11.7474 31.3804C11.6241 31.1925 11.4888 31.0801 11.3948 31.0177C11.5132 30.8719 11.5836 30.6756 11.5836 30.4817C11.5836 30.0383 11.2269 29.6777 10.7892 29.6777C10.3871 29.6777 10.0433 30.01 10.0023 30.4214C9.91874 30.4313 9.78219 30.451 9.64953 30.4822L11.7767 27.9807C11.798 27.9921 11.8216 27.9967 11.8439 28.0051L11.7474 31.3804ZM12.0877 31.715L11.9089 32.1782L12.0265 28.052C12.0399 28.0536 12.0534 28.0552 12.0669 28.0552C12.1106 28.0552 12.1505 28.0422 12.1918 28.0339L12.0877 31.715ZM12.2858 31.2031L12.3772 27.9625C12.4011 27.9479 12.4283 27.9396 12.4504 27.923C12.4948 28.1922 12.7092 28.4137 12.9919 28.45C13.0171 28.4531 13.0436 28.4547 13.0688 28.4547C13.1866 28.4547 13.2918 28.4129 13.3873 28.355L12.2858 31.2031ZM13.5124 27.8781C13.4818 28.1229 13.2575 28.3041 13.014 28.2714C12.7697 28.2403 12.595 28.0128 12.6254 27.7657C12.6524 27.5391 12.8429 27.3688 13.0682 27.3688C13.0864 27.3688 13.1054 27.3703 13.1235 27.3727C13.3683 27.4028 13.5428 27.6302 13.5124 27.8781ZM13.6651 27.6362C13.5942 27.4056 13.3982 27.2255 13.1459 27.1935C12.974 27.173 12.8107 27.2231 12.6856 27.3205C12.6744 27.2558 12.6536 27.1951 12.6248 27.1372L14.6858 24.9963L13.6651 27.6362ZM14.8566 24.5536L12.5158 26.9851C12.4867 26.9547 12.4517 26.9311 12.4161 26.9059L15.0057 24.1686L14.8566 24.5536ZM13.9182 24.3369C13.8598 24.3369 13.8021 24.3338 13.745 24.3262C13.8587 24.2624 13.9501 24.1588 13.9974 24.0311C14.0568 23.8698 14.0384 23.6941 13.9579 23.5518C14.1053 23.5443 14.2429 23.4895 14.3543 23.4013C14.392 23.5838 14.5181 23.745 14.7037 23.8158C14.7673 23.8403 14.8346 23.8525 14.9015 23.8525C14.9407 23.8525 14.9797 23.8486 15.0181 23.8403C14.7414 24.1489 14.3434 24.3369 13.9182 24.3369ZM15.2507 23.4228C15.1801 23.6128 14.96 23.718 14.7689 23.6443C14.577 23.5713 14.4784 23.3545 14.5487 23.1598C14.5571 23.1424 14.5648 23.1248 14.5724 23.1074C14.6588 22.9462 14.8585 22.8693 15.0319 22.9355C15.2263 23.0067 15.3239 23.2258 15.2507 23.4228ZM15.3865 23.0121C15.3252 22.9002 15.2242 22.8107 15.0973 22.7626C15.0337 22.739 14.9672 22.7268 14.9003 22.7268C14.8006 22.7268 14.7055 22.7528 14.6225 22.8013C14.6139 22.6554 14.5612 22.5147 14.4719 22.4005C14.6575 22.3618 14.8159 22.2294 14.8836 22.0461C14.9223 21.9413 14.9278 21.8317 14.906 21.7297C15.2577 22.0493 15.4413 22.529 15.3865 23.0121ZM15.4345 22.1585C15.3107 21.8787 15.1092 21.636 14.8504 21.4574C14.9 21.4392 14.9478 21.4156 14.9898 21.3813C15.0994 21.2977 15.1697 21.1744 15.1866 21.0368C15.2206 20.7546 15.0217 20.4955 14.7414 20.4589H14.7411C14.6069 20.4436 14.4727 20.4787 14.3647 20.5633C14.2557 20.6479 14.1861 20.7712 14.1692 20.9088C14.1562 21.0116 14.1827 21.1066 14.2242 21.1926C14.1238 21.1736 14.0215 21.163 13.9179 21.163C13.7834 21.163 13.6531 21.1835 13.5259 21.2154L13.5729 21.1567C13.8221 21.1406 14.0428 21.0988 14.0521 21.0973C14.095 21.089 14.1269 21.0508 14.1269 21.0075C14.1269 20.698 14.3722 20.4374 14.6635 20.4374C14.9623 20.4374 15.2058 20.6835 15.2058 20.9854C15.2058 21.1528 15.1541 21.2728 15.0391 21.3725C15.0179 21.3906 15.0064 21.4174 15.0075 21.4462C15.009 21.4735 15.0228 21.4994 15.0454 21.5163C15.048 21.5186 15.2907 21.6972 15.4332 22.0731C15.4345 22.0755 15.4358 22.0778 15.4363 22.0799L15.4345 22.1585ZM15.4478 21.697C15.3696 21.5776 15.2902 21.4908 15.2338 21.4368C15.3377 21.3151 15.3891 21.1661 15.3891 20.9859C15.3891 20.5822 15.0638 20.2549 14.663 20.2549C14.2985 20.2549 13.9857 20.5563 13.947 20.9298C13.8943 20.9382 13.818 20.9496 13.7305 20.9594L15.2004 19.1121C15.2813 19.1767 15.38 19.2196 15.4895 19.2339C15.501 19.2349 15.5111 19.2362 15.5223 19.237L15.4478 21.697ZM15.7268 22.3047L15.6035 22.6212L15.7071 19.2198C15.7497 19.2087 15.7917 19.1934 15.8314 19.1726L15.7268 22.3047ZM15.8091 18.9698C15.7242 19.0366 15.6188 19.0636 15.5137 19.0506C15.293 19.0231 15.136 18.8193 15.1637 18.5958C15.177 18.4862 15.2318 18.3894 15.318 18.3219C15.3893 18.2668 15.4748 18.2372 15.5628 18.2372C15.5789 18.2372 15.5957 18.2383 15.6113 18.2404H15.6131C15.6617 18.2469 15.7074 18.2609 15.7484 18.2835C15.7549 18.2886 15.7627 18.2928 15.7702 18.2959C15.8335 18.3346 15.8852 18.3899 15.9197 18.4556C15.9176 18.4828 15.9278 18.5111 15.9496 18.5303C15.9657 18.583 15.9701 18.6389 15.9633 18.6962C15.9503 18.8055 15.8958 18.9026 15.8091 18.9698ZM15.928 21.7837L16.009 19.3554C16.1009 19.4862 16.2453 19.5797 16.4145 19.6012C16.5589 19.6204 16.6941 19.5797 16.8107 19.5002L15.928 21.7837ZM16.8873 19.0641C16.8735 19.1736 16.8187 19.271 16.7325 19.338C16.6484 19.4036 16.5399 19.4335 16.4379 19.4192C16.2164 19.3909 16.0602 19.1861 16.0872 18.9628C16.1009 18.8533 16.1554 18.7565 16.2421 18.689C16.313 18.6337 16.3989 18.6046 16.4864 18.6046C16.5025 18.6046 16.5194 18.6059 16.5355 18.6077C16.7569 18.6365 16.9142 18.8416 16.8873 19.0641ZM17.0531 18.8735C16.9958 18.6425 16.8073 18.4576 16.5599 18.4262H16.5594C16.4085 18.4081 16.2619 18.4475 16.1406 18.5368C16.1365 18.515 16.1315 18.4932 16.1245 18.4714L18.1367 16.0735L17.0531 18.8735ZM18.341 15.5426L16.0345 18.2928C16.0069 18.2554 15.975 18.2214 15.94 18.1916L18.0991 15.6049C18.2172 15.5306 18.3247 15.4418 18.4194 15.3388L18.341 15.5426ZM18.5246 14.8076C18.4591 14.9836 18.2522 15.0805 18.0786 15.013C17.9013 14.946 17.8107 14.7464 17.8732 14.5667C17.8826 14.5486 17.8906 14.5296 17.8979 14.5109C17.9804 14.364 18.1645 14.2936 18.3223 14.3544C18.5022 14.4229 18.5926 14.6257 18.5246 14.8076ZM18.6549 14.4071C18.5952 14.3074 18.5033 14.2269 18.3875 14.1828C18.327 14.1599 18.2639 14.1485 18.2001 14.1485C18.1087 14.1485 18.0209 14.1721 17.9441 14.2144C17.9345 14.0826 17.8872 13.9559 17.8078 13.8505C17.9799 13.8118 18.1263 13.6872 18.1907 13.5164C18.2252 13.4221 18.2325 13.3232 18.2156 13.2308C18.5334 13.5278 18.6998 13.9652 18.6549 14.4071ZM18.8428 14.1623C18.8423 14.1659 18.8408 14.1688 18.8405 14.1721C18.8057 13.5649 18.427 13.0047 17.8291 12.7775C17.6507 12.71 17.4649 12.6755 17.2767 12.6755C16.8909 12.6755 16.5293 12.8235 16.2486 13.0693L16.0962 13.1576C16.3179 12.6316 16.1692 12.5377 16.0552 12.4678C15.9078 12.377 15.7756 12.4839 15.6855 12.5338C15.5947 12.5828 15.0599 13.5595 15.0599 13.5595L15.0895 12.9754L15.3961 12.4603C15.4122 12.4325 15.4197 12.4024 15.4189 12.3726C15.7567 12.0221 16.4301 11.9346 17.4238 12.113C18.3735 12.2812 19.0108 13.201 18.8428 14.1623Z"
+                      fill="#A39161"
+                    ></path>
+                  </g>
+                  <defs>
+                    <clipPath id="clip0_1489_709">
+                      <rect
+                        width="20"
+                        height="31.2"
+                        fill="currentColor"
+                        transform="translate(0 4.8)"
+                      ></rect>
+                    </clipPath>
+                  </defs>
+                </svg>
+                <span className="govie-visually-hidden">gov.ie</span>
+              </a>
+            </div>
+            <div className="govie-!-font-size-24">
+              <strong>{t("myLifeEvents")}</strong>
+            </div>
+          </div>
+          <div
+            style={{
+              display: "flex",
+              justifyContent: "space-between",
+              alignItems: "center",
+              gap: "20px",
+              marginTop: "3px",
+            }}
+          >
+            <div style={{ display: "flex", alignItems: "center", gap: "10px" }}>
+              <Link
+                className={`govie-link govie-link--inverse govie-link--no-underline ${
+                  pathSlice.at(1) === "en" ? "govie-!-font-weight-bold" : ""
+                }`.trim()}
+                href={new URL("/en/" + path, process.env.HOST_URL).href}
+              >
+                English
+              </Link>
+              <div
+                style={{
+                  height: "14px",
+                  width: "1px",
+                  borderLeft: `1px solid ${ds.colours.ogcio.white}`,
+                }}
+              />
+
+              <Link
+                className={`govie-link govie-link--inverse govie-link--no-underline  ${
+                  pathSlice.at(1) === "ga" ? "govie-!-font-weight-bold" : ""
+                }`.trim()}
+                href={new URL("/ga/" + path, process.env.HOST_URL).href}
+              >
+                Gealic
+              </Link>
+            </div>
+            <NotificationsLink />
+            <UserIcon />
+
+            <Link href="/logout" style={{ display: "flex" }}>
+              <ds.Icon icon="logout" color={ds.colours.ogcio.white} size={22} />
+            </Link>
+          </div>
+        </div>
+      </div>
+    </header>
+  );
+};
